+++ src/utils/transformingDictionaries/createStrokeHintForPhrase.test.ts
@@ -2,2368 +2,11 @@ import chooseOutlineForPhrase from "./chooseOutlineForPhrase";
 import createStrokeHintForPhrase from "./createStrokeHintForPhrase";
 import createAGlobalLookupDictionary from "./createAGlobalLookupDictionary";
 import { AffixList } from "../affixList";
-import { PersonalDictionaryNameAndContents } from "../../types";
-
-const testTypeyTypeDict = {
-  "TK-LS": "{^^}",
-  "KP-PL": "example",
-  "TR*PL": "{^}™",
-  "SPWO*L/TRAEUD/PHARBG": "{^}™",
-  "PWRABG": "<{^}",
-  "HR*PB": "<",
-  "HAERB": "#{^}",
-  "TPHRORB": "$",
-  "H-PBZ": "--{^}",
-  "H-PBS": "-{^}",
-  "H*B": "-",
-  "H-PB": "{^-^}",
-  "TPR-BGT": "\\{{^}",
-  "HAOEPT": "{>}http://{^}",
-  "P*ERS": "{&%}",
-  "1-BGS": "{^}{#F1}{^}",
-  "KHR-PB": "{^}:{^}",
-  "KR-RT": "{^}^{^}",
-  "P*PB": "{^}({^}",
-  "TP-P/TP-P": "{~|‘^}",
-  "TP-L/TP-L": "{^~|’}",
-  "KW-GS/KW-GS": "{~|“^}",
-  "KR-GS/KR-GS": "{^~|”}",
-  "AEU": "a",
-  "EU": "I",
-  "S-P": "{^ ^}",
-  "H-F": "{?}",
-  "KWEZ": "?",
-  "KW-BG": "{,}",
-  "TP-PL": "{.}",
-  "PR-D": ".",
-  "TOPL": "Tom",
-  "H*ET/*ER": "heather",
-  "TAOUZ": "Tuesday",
-  "TPEUFRT": "first",
-  "A/TKRES": "address",
-  "PWED": "bed",
-  "PWED KW-BG": "bed,",
-  "PHAPB": "man",
-  "SKHRAPL": "{!}",
-  "STKPWEPBG": "and again",
-  "SKP": "and",
-  "TKPWEPB": "again",
-  "PHO*EUD": "media",
-  "KWAOER/REU": "query",
-  "STPHAT/RA": "Sinatra",
-  "AE": "{^'}",
-  "PURB": "push",
-  "O*RPBLG": "origin",
-  "PHAFRT": "master",
-  "TKEUF": "diff",
-  "TK*RB": "{--}",
-  "KAERBD": "cached",
-  "A*EPBG": "{^>^}",
-  "AEPBG": "{^<^}",
-  "OEU": "{^/^}",
-  "TAOEULT": "title",
-  "HRERPB": "learn",
-  "OERBGS": "oh{,}",
-  "HRAO*EUBG": "{,}like{,}",
-  "HREPBT": "lent",
-  "SKROL": "scroll",
-  "TPAR/PHER": "farmer",
-  "KR-GS": "{^~|\"}",
-  "KW-GS": "{~|\"^}",
-  "T": "it",
-  "-B": "be",
-  "PWE": "{be^}",
-  "KWAS/KWREU": "{quasi-^}",
-  "KWA/SEU": "quasi",
-  "SPAOERL": "experimental",
-  "TKPWHRAOEU": "{gly-^}",
-  "KPAOEUD": "oxide",
-  "EBGT/PHEU": "{^ectomy}",
-  "SAEUD": "said",
-  "KPAOUR": "computer",
-  "KAT": "cat",
-  "KET/*L": "kettle",
-  "K": "can",
-  "K-PBT": "can't",
-  "HO*UFS": "houses",
-  "HER": "her",
-  "HROPBG": "long",
-  "TPHAR/AEUT": "narrate",
-  "SAO*ET": "seethe",
-  "PWEUPBG": "bing",
-  "PWEUPB/-PBLG": "binge",
-  "KUF": "cuff",
-  "U": "you",
-  "STOR": "store",
-  "RAOPL": "room",
-  "RAO*PL": "{^room}",
-  "OUDZ": "outside",
-  "HEUT": "hit",
-  "PHEUS": "miss",
-  "H-PLS": "hit-and-miss",
-  "PWUF/ET": "buffet",
-  "WAPBGD": "wandering", // currently pre-sorted to best stroke first
-  "WAPB/TKER/-G": "wandering",
-  "HROPBLG": "lodge",
-  "S-PBT": "isn't",
-  "PHAEUD/*EPB": "maiden",
-  "PHAEUD/EPB": "maiden",
-  "1-9DZ": "$100",
-  "1-DZ": "$100",
-  "2-DZ": "$200",
-  "3-DZ": "$300",
-  "4-DZ": "$400",
-  "#45/TK-PL": "$45",
-  "45/TK-PL": "$45",
-  "5DZ": "$500",
-  "-6DZ": "$600",
-  "-7DZ": "$700",
-  "0EU": "0",
-  "#-Z": "00",
-  "10*EU": "01",
-  "10EU": "01",
-  "20EU": "02",
-  "30EU": "03",
-  "40EU": "04",
-  "50EU": "05",
-  "1/TPW-S": "1 tablespoon",
-  "#S/W-B/THUZ": "1,000",
-  "1/THO*EUPB": "1,000",
-  "1/THOEUB": "1,000",
-  "TPHOUZ": "1,000",
-  "HA*F": "1/2",
-  "TPHA*F": "1/2",
-  "130EU": "1/3",
-  "TH*EURD": "1/3",
-  "140EU": "1/4",
-  "KWA*RT": "1/4",
-  "10EU8": "1/8",
-  "10/PERS/APBLG": "10 percentage",
-  "1/0": "10",
-  "10/P*ERS": "10%",
-  "#SO/W-B/THUZ": "10,000",
-  "1-Z": "100",
-  "TPHUPBZ": "100",
-  "1/THOUZ": "1000",
-  "1-DZ/1": "1001",
-  "10/1": "101",
-  "10/*PBT": "10th",
-  "10/*T": "10th",
-  "1-D": "11",
-  "1BGD": "11:00",
-  "1-D/*T": "11th",
-  "1/2": "12",
-  "12/TPH/1": "12-in-1",
-  "12/1": "121",
-  "12R5": "125",
-  "12-BG": "12:00",
-  "12/*T": "12th",
-  "1/3": "13",
-  "14": "14",
-  "14*": "14",
-  "1/5": "15",
-  "1/6": "16",
-  "1/7": "17",
-  "1/8": "18",
-  "18/50/-S": "1850s",
-  "1-9": "19",
-  "1-9D": "19",
-  "19/30/-S": "1930s",
-  "1-9/50": "1950",
-  "1-9D/50": "1950",
-  "TPHEFPBT": "197",
-  "19/OEUP": "1970",
-  "TPHEFPBD": "1970",
-  "TPHAEUPBT": "198",
-  "TPHAEUPBD": "1980",
-  "19/8/6": "1986",
-  "TPHEUPBD": "1990",
-  "1-BG": "1:00",
-  "1/PW*": "1b",
-  "1/S*/T*": "1st",
-  "2/KP*/4": "2 x 4",
-  "#T-/W-B/THUZ": "2,000",
-  "230EU": "2/3",
-  "2/0": "20",
-  "2-Z": "200",
-  "KWRAOUPB/1": "2001",
-  "STWOUPB/1": "2001",
-  "TWOUPB/-P": "2007",
-  "TWOUPB/7": "2007",
-  "TWOUPB/10": "2010",
-  "20/-S": "20s",
-  "12EU": "21",
-  "2-D": "22",
-  "2/3": "23",
-  "2/4": "24",
-  "#240": "240",
-  "2/5": "25",
-  "2/6": "26",
-  "2/6/0": "260",
-  "2/7": "27",
-  "2/8": "28",
-  "2/9": "29",
-  "2-BG": "2:00",
-  "2/TK*": "2d",
-  "2/*PBD": "2nd",
-  "2-S": "2s",
-  "3/T-BS": "3 tablespoon",
-  "3/T-BS/-S": "3 tablespoons",
-  "3/KP*/6": "3 x 6",
-  "#P-/W-B/THUZ": "3,000",
-  "340EU": "3/4",
-  "30EU8": "3/8",
-  "3/0": "30",
-  "3-Z": "300",
-  "30*S": "30s",
-  "30/-S": "30s",
-  "13EU": "31",
-  "23EU": "32",
-  "3-D": "33",
-  "3/4": "34",
-  "3/5": "35",
-  "PAO/TKEG": "350 degree",
-  "3/6": "36",
-  "36/0/TKEG/-S": "360 degrees",
-  "3/7": "37",
-  "3/8": "38",
-  "3/9": "39",
-  "3-BG": "3:00",
-  "30*EUD": "3D",
-  "3/R*D": "3rd",
-  "4/T-BS/-S": "4 tablespoons",
-  "#H/W-B/THUZ": "4,000",
-  "40": "40",
-  "40*": "40",
-  "4-Z": "400",
-  "40ES": "40s",
-  "40S": "40s",
-  "40RBGS": "40{,}",
-  "14EU": "41",
-  "24EU": "42",
-  "34EU": "43",
-  "4-D": "44",
-  "4/5": "45",
-  "4/6": "46",
-  "4/7": "47",
-  "4/8": "48",
-  "4/9": "49",
-  "4-BG": "4:00",
-  "4-S": "4s",
-  "4/*T": "4th",
-  "R5": "5",
-  "TPEUF/P*ERS": "5%",
-  "#A/W-B/THUZ": "5,000",
-  "5/-P/4": "5.4",
-  "50EU8": "5/8",
-  "50/KREPBT/-S": "50 cents",
-  "5/0": "50",
-  "5Z": "500",
-  "50S": "50s",
-  "15EU": "51",
-  "25*EU": "52",
-  "25EU": "52",
-  "35EU": "53",
-  "45EU": "54",
-  "5D": "55",
-  "56": "56",
-  "57": "57",
-  "58": "58",
-  "59": "59",
-  "5BG": "5:00",
-  "5/*T": "5th",
-  "#F/W-B/THUZ": "6,000",
-  "0EU6": "60",
-  "-6Z": "600",
-  "1EU6": "61",
-  "2EU6": "62",
-  "3EU6": "63",
-  "4EU6": "64",
-  "5EU6": "65",
-  "-6D": "66",
-  "67": "67",
-  "68": "68",
-  "69": "69",
-  "K-6": "6:00",
-  "6/*T": "6th",
-  "#-P/W-B/THUZ": "7,000",
-  "0EU7": "70",
-  "-7Z": "700",
-  "0EU7S": "70s",
-  "1EU7": "71",
-  "2EU7": "72",
-  "3EU7": "73",
-  "4EU7": "74",
-  "5EU7": "75",
-  "EU67": "76",
-  "-7D": "77",
-  "78": "78",
-  "79": "79",
-  "K-7": "7:00",
-  "7/*T": "7th",
-  "#L/W-B/THUZ": "8,000",
-  "0EU8/KREPBT/SKP-S": "80 cents and",
-  "0EU8": "80",
-  "-8Z": "800",
-  "0EU8S": "80s",
-  "1EU8": "81",
-  "2EU8": "82",
-  "3EU8": "83",
-  "4EU8": "84",
-  "8/5/KREPBT/-S": "85 cents",
-  "5EU8": "85",
-  "EU68": "86",
-  "EU78": "87",
-  "-8D": "88",
-  "89": "89",
-  "K-8": "8:00",
-  "8/*T": "8th",
-  "8/SR*/O*": "8vo",
-  "#-T/W-B/THUZ": "9,000",
-  "0EU9": "90",
-  "0EU9/PERS": "90%",
-  "-9Z": "900",
-  "EU9": "900",
-  "1EU9": "91",
-  "2EU9": "92",
-  "3EU9": "93",
-  "4EU9": "94",
-  "5EU9": "95",
-  "EU69": "96",
-  "EU79": "97",
-  "EU89": "98",
-  "-9D": "99",
-  "K-9": "9:00",
-  "9/*T": "9th",
-  "1-R": "I",
-  "2-R": "II",
-  "3-R": "III",
-  "4-R": "IV",
-  "R-9": "IX",
-  "R-6": "VI",
-  "R-7": "VII",
-  "R-8": "VIII",
-  "10R": "X",
-  "1-RD": "XI",
-  "12-R": "XII",
-  "0Z": "{&00}",
-  "#O": "{&0}",
-  "0RBGS": "{&0}",
-  "#S": "{&1}",
-  "1-RBGS": "{&1}",
-  "#T-": "{&2}",
-  "2-RBGS": "{&2}",
-  "#P-": "{&3}",
-  "3-RBGS": "{&3}",
-  "#H": "{&4}",
-  "4-RBGS": "{&4}",
-  "#A": "{&5}",
-  "5RBGS": "{&5}",
-  "#F": "{&6}",
-  "1KWR-6": "{&6}",
-  "#-P": "{&7}",
-  "1KWR-7": "{&7}",
-  "#L": "{&8}",
-  "1KWR-8": "{&8}",
-  "#-T": "{&9}",
-  "1KWR-9": "{&9}",
-  "TWOUPB": "{200^}",
-  "30*U": "{:}30",
-  "40U": "{:}40",
-  "45*U": "{:}45",
-  "0RZ": "{^0rz}",
-  "50*EUBG": "{^:05}",
-  "50EUBG": "{^:05}",
-  "10*BG": "{^:10}",
-  "10BG": "{^:10}",
-  "15*BG": "{^:15}",
-  "15BG": "{^:15}",
-  "20*BG": "{^:20}",
-  "20BG": "{^:20}",
-  "25*BG": "{^:25}",
-  "25BG": "{^:25}",
-  "30*BG": "{^:30}",
-  "30BG": "{^:30}",
-  "35*BG": "{^:35}",
-  "35BG": "{^:35}",
-  "40*BG": "{^:40}",
-  "40BG": "{^:40}",
-  "45*BG": "{^:45}",
-  "45BG": "{^:45}",
-  "50*BG": "{^:50}",
-  "50BG": "{^:50}",
-  "5*BGD": "{^:55}",
-  "5BGD": "{^:55}",
-  "TWOUPB/9/TKHRAR/-S": "2009 dollars",
-  "TWOUPBD": "2000",
-  "TWOUZ": "2000",
-  "KPR-T": "©",
-  "TPH": "in",
-  "KWROUR": "your",
-  "KR-PGS": "cross-petition",
-  "KROPL": "{^.com}",
-  "TKR-FPLT": "Dr.",
-  "TPHEL/KWREU": "nelly",
-  "KOFR": "cover",
-  "PHAS": "mass",
-  "AUFR/-T": "{over-the-^}",
-  "AEUPB/SKWREU": "{ani-^}",
-  "A*EURBG": "{^-acre}",
-  "*EUPB/HRAU": "{^-in-law}",
-  "STOEUPB": "steno",
-  "HRAOEUF": "life",
-  "-T": "the",
-  "-F": "of",
-  // "SKP": "and",
-  "TO": "to",
-  // "TPH": "in",
-  // "EU": "I",
-  "THA": "that",
-  "WAS": "was",
-  "HEUS": "his",
-  "HE": "he",
-  // "T": "it",
-  "W": "with",
-  "S": "is",
-  "TPOR": "for",
-  "AZ": "as",
-  "H": "had",
-  // "U": "you",
-  "TPHOT": "not",
-  // "-B": "be",
-  // "HER": "her",
-  "OPB": "on",
-  "AT": "at",
-  "PWEU": "by",
-  "WEU": "which",
-  "SR": "have",
-  "OR": "or",
-  "TPR": "from",
-  "TH": "this",
-  "HEUPL": "him",
-  "PWUT": "but",
-  "AUL": "all",
-  "SHE": "she",
-  "THE": "they",
-  "WR": "were",
-  "PHEU": "my",
-  "R": "are",
-  "PHE": "me",
-  "WUPB": "one",
-  "THAEUR": "their",
-  "SO": "so",
-  "APB": "an",
-  "SED": "said",
-  "THEPL": "them",
-  "WE": "we",
-  "WHO": "who",
-  "WO": "would",
-  "PW-PB": "been",
-  "HR": "will",
-  "TPHO": "no",
-  "WHEPB": "when",
-  "THR": "there",
-  "TP": "if",
-  "PHOR": "more",
-  "OUT": "out",
-  "UP": "up",
-  "TPHAO": "into",
-  "TKO": "do",
-  "TPHEU": "any",
-  // "KWROUR": "your",
-  "WHA": "what",
-  "HAS": "has",
-  // "PHAPB": "man",
-  "KO": "could",
-  "OER": "other",
-  "THAPB": "than",
-  "OUR": "our",
-  "SOPL": "some",
-  "SRE": "very",
-  "TAOEUPL": "time",
-  "POPB": "upon",
-  "PW": "about",
-  "PHAE": "may",
-  "EUTS": "its",
-  "OEPBL": "only",
-  "TPHOU": "now",
-  "HRAOEUBG": "like",
-  "HREUL": "little",
-  "THEPB": "then",
-  // "K": "can",
-  "SHO": "should",
-  "PHAED": "made",
-  "TK": "did",
-  "US": "us",
-  "SUFP": "such",
-  // "AEU": "a",
-  "TKPWRAET": "great",
-  "PW-FR": "before",
-  "PHUFT": "must",
-  "TWO": "two",
-  "THEZ": "these",
-  "SAOE": "see",
-  "TPHOE": "know",
-  "OEFR": "over",
-  "PHUFP": "much",
-  "TKOUPB": "down",
-  "AF": "after",
-  // "TPEUFRT": "first",
-  "PHR-FPLT": "Mr.",
-  "TKPWAOD": "good",
-  "PHEPB": "men",
-};
-
-const testPloverDict = {
-  "A/HREF": "<a href=\"{^}",
-  "APBD": "and",
-  "SKP": "and",
-  "SP": "and",
-  "P-R": "for",
-  "PHO": "no",
-  "WH": "when",
-  "SPH": "some",
-  "TPEURS": "first",
-  "KR-PGS": "cross-petition",
-  "TKR*FPLT": "Dr.{-|}",
-  "PH-BGS": "Mx.{-|}",
-  "PHRARPLS": "Mr. and Mrs.",
-  "KHAPBT": "chant",
-  "EL/TKREUPBLG": "Eldridge",
-  "TWA*S": "{~|'^}twas",
-  "SKWR*EPL": "gentlemen",
-  "OP/TOPL/TREUFT": "optometrist",
-  "POED/TREUFT": "podiatrist",
-};
-
-const testEmojiDict = {
-  "PHOEPBLG/T*/PH*": "™",
-};
-const testRubyDict = {
-  "TPHRARB/PWR-BGT": "flash[:{^}",
-};
-const testReactDict = {
-  "O*B/P-P": "Object.{^}",
-};
-
-// Some prefix and suffix entries are commented out because they are alternative strokes for prefix/suffix translations and the preferred stroke already exists.
-// Keeping entries uncommented out improves the chances of finding a valid dictionary entry. Previously, we would sort the affix arrays to ensure "preferred" strokes were used where possible, but we won't have that control with personal dictionaries.
-const testPrefixesDict = {
-  "*EBGS/TRA": "{extra^}",
-  "*EFR": "{every^}",
-  "*ERBGS": "{extra^}",
-  "*ERT/RO": "{erythro^}",
-  "*ERTD/RO": "{erythro^}",
-  "*EUG": "{Ig^}",
-  "*EUL/KWRO": "{ileo^}",
-  "*EUPBG/TKO": "{incudo^}",
-  "*EUPL": "{im^}",
-  "*FR": "{ever^}",
-  "-T/TPHAEUS/KWRO": "{of the naso^}",
-  "A": "{a^}",
-  "A*ED": "{ad^}",
-  "A*EPBT": "{ante^}",
-  "A*ER": "{aero^}",
-  "A*EUPBT": "{anti^}",
-  "A*EUPBT/H-PB": "{anti-^}",
-  "A*F": "{after^}",
-  "A*LG/*L": "{alkyl^}",
-  "A*PBT/RA": "{anthra^}",
-  "A*PBT/RO": "{anthro^}",
-  "A*RT/RO": "{arthro^}",
-  "A*T/RO": "{athero^}",
-  "A*T/ROE": "{athero^}",
-  "A*UT": "{auto^}",
-  "A/PHAOEB": "{ameb^}",
-  "A/PHEUG/TKA/HRO": "{amygdalo^}",
-  "A/PHO*EUPB": "{amino^}",
-  "A/PO": "{apo^}",
-  "A/SAOET/*L": "{acetyl^}",
-  "A/SAOET/KWRO": "{aceto^}",
-  "A/THAOER/SKWRO": "{athero^}",
-  "A/TKAOEPB/SKWRO": "{adeno^}",
-  "A/TKRAOEPB/SKWRO": "{adreno^}",
-  "A/TPHA": "{ana^}",
-  "A/TPHAFL": "{anaphylo^}",
-  "AB": "{ab^}",
-  "ABG/RO": "{acro^}",
-  "ABG/RO/PHEU/KWRO": "{acromio^}",
-  "ABGS/KWRO": "{axio^}",
-  "ABGS/SKWRO": "{axo^}",
-  "AD/PO": "{adipo^}",
-  "AEPBT": "{ante^}",
-  "AEPL": "{ambi^}",
-  "AEU/HREUPBG/WO": "{a linguo^}",
-  "AEU/KOS": "{eicosa^}",
-  "AEU/KPAOEPB/KWRO": "{a xeno^}",
-  "AEUR/KWRE": "{arye^}",
-  "AFR": "{after^}",
-  "AL/HRO": "{allo^}",
-  "AO*EP": "{epi^}",
-  "AO*UPB": "{uni^}",
-  "AOE": "{e^}",
-  "AOE/HREBG/TRO": "{electro^}",
-  "AOE/HREUPT/KWRO": "{elipto^}",
-  "AOE/PHET/SKWRO": "{emeto^}",
-  "AOEBG/SKWRO": "{eco^}",
-  "AOET/KWRO": "{etio^}",
-  "AOET/KWROE": "{etio^}",
-  "AOEUD/SKWRO": "{iodo^}",
-  "AOEUS/KWRO": "{iso^}",
-  "AOEUS/SKWRO": "{iso^}",
-  "AOU/TREUBG/HRO": "{utriculo^}",
-  "AOUB/*ER": "{uber^}",
-  "AOUF/KWROE": "{uveo^}",
-  "AOUPB": "{uni^}",
-  "AOUPB/SKWREU": "{uni^}",
-  "AOUT": "{out^}",
-  "AOUT/RO": "{utero^}",
-  "AP/KWRO": "{apo^}",
-  "APB/AF/HRO": "{anaphylo^}",
-  "APB/APBT/RO": "{an antero^}",
-  "APB/SKWRA": "{ana^}",
-  "APB/TE": "{ante^}",
-  "APB/TER/KWRO": "{antero^}",
-  "APB/THRA": "{anthra^}",
-  "APB/THRO/PO": "{anthropo^}",
-  "APBG/KWRO": "{angio^}",
-  "APBG/KWROE": "{angio^}",
-  "APBG/SKWRO": "{angio^}",
-  "APBGS/KWRO": "{anxio^}",
-  "APBT/RO": "{antero^}",
-  "APBT/SKWR*E": "{ante^}",
-  "APBT/SKWRE": "{ante^}",
-  "APL/HRO": "{amelo^}",
-  "AR/KWREU": "{ary^}",
-  "AR/TAOER/KWRO": "{arterio^}",
-  "AR/TAOER/KWROE": "{arterio^}",
-  "AR/TAOER/SKWRO": "{arterio^}",
-  "ARBG/KWROE": "{archeo^}",
-  "ARPBD/SKWRO": "{andro^}",
-  "ART/RO": "{arterio^}",
-  "AS/TPHAOE/TOE": "{acineto^}",
-  "AUF": "{off^}",
-  "AUFR": "{over^}",
-  "AUFT/RO": "{Austro^}",
-  "AUP": "{up^}",
-  "AUPB": "{on^}",
-  "AUR": "{or^}",
-  "AUR/EUBG/HRO": "{auriculo^}",
-  "AURBG/HRO": "{auriculo^}",
-  "EBG/KWRO": "{echo^}",
-  "EBGS/AOEUT/SKWRO": "{excito^}",
-  "EBGS/AOEUT/SKWRO*": "{excito^}",
-  "EBGS/OE": "{exo^}",
-  "EBGS/SKWRO": "{exo^}",
-  "EBGT/SKWRO": "{ecto^}",
-  "EP": "{epi^}",
-  "EP/SKWREU": "{epi^}",
-  "EPB": "{en^}",
-  "EPB/SEF/HRO": "{encephalo^}",
-  "EPB/SEFL": "{encephalo^}",
-  "EPB/SEFL/HRO": "{encephalo^}",
-  "EPB/TER/SKWRO": "{entero^}",
-  "EPB/TKO": "{endo^}",
-  "EPBD/SKWRO": "{endo^}",
-  "EPBT/*ER/SKWRO": "{entero^}",
-  "EPBT/RO": "{entero^}",
-  "EPBT/SKWRO": "{ento^}",
-  "EPL": "{em^}",
-  "ER": "{er^}",
-  "ERBGS": "{extra^}",
-  "EUD/KWRO": "{idio^}",
-  "EUD/KWROE": "{idio^}",
-  "EUD/SKWRO": "{ideo^}",
-  "EUFBG/KWRO": "{ischio^}",
-  "EUFRPB": "{infra^}",
-  "EUL/KWRO": "{ilio^}",
-  "EUL/KWRO*E": "{ileo^}",
-  "EUL/KWROE": "{ilio^}",
-  "EUPB": "{in^}",
-  "EUPB/PHAOUPB/SKWRO": "{immuno^}",
-  "EUPB/PHAOUPB/TPHOE": "{immuno^}",
-  "EUPB/TPRA": "{infra^}",
-  "EUPB/TRA": "{intra^}",
-  "EUPBS/HRO": "{insulo^}",
-  "EUPBT": "{inter^}",
-  "EUPL": "{im^}",
-  "EUPL/PHAOUPB/KWRO": "{immuno^}",
-  "EUPL/PHAOUPB/SKWRO": "{immuno^}",
-  "EUPL/PHAOUPB/TPHO": "{immuno^}",
-  "EUR": "{ir^}",
-  "EURD/SKWRO": "{irido^}",
-  "EURPBT": "{intro^}",
-  "H*ERT": "{hetero^}",
-  "HA*EPL": "{hemi^}",
-  "HA*EPT": "{hepato^}",
-  "HAEPT/SKWRO": "{hepato^}",
-  "HAO*EUP": "{hypo^}",
-  "HAO*EURD": "{hydro^}",
-  "HAO*RD": "{hydro^}",
-  "HAOEL/KWRO": "{helio^}",
-  "HAOEPL": "{hemo^}",
-  "HAOEPL/KWRA": "{hemia^}",
-  "HAOEPL/SKWRO": "{hemo^}",
-  "HAOEPLT/SKWRO": "{hemato^}",
-  "HAOEPLT/TO": "{hemato^}",
-  "HAOEU/POE": "{hypo^}",
-  "HAOEU/SKWRO": "{hyo^}",
-  "HAOEU/TKROBGS/KWREU": "{hydroxy^}",
-  "HAOEU/TKROE": "{hydro^}",
-  "HAOEUB/SKWRO": "{ribo^}",
-  "HAOEUBG/SKWRO": "{myco^}",
-  "HAOEUD/RO": "{hydro^}",
-  "HAOEUD/ROBGS/KWREU": "{hydroxy^}",
-  "HAOEUD/TKROE": "{hydro^}",
-  "HAOEUG/ROE": "{hygro^}",
-  "HAOEUP/KWRO": "{hypo^}",
-  "HAOEUP/O*": "{hypo^}",
-  "HAOEUP/OE": "{hypo^}",
-  "HAOEUP/SKWRO": "{hypo^}",
-  "HAOEURD": "{hydro^}",
-  "HAOEURD/OBGS/KWREU": "{hydroxy^}",
-  "HAOEURD/OBGS/SKWREU": "{hydroxy^}",
-  "HAOEURD/RA": "{hydra^}",
-  "HAOEURD/SKWRO": "{hydro^}",
-  "HAOEURP": "{hyper^}",
-  "HAPT/SKWRO": "{hapto^}",
-  "HE/PA/TO": "{hepato^}",
-  "HEL/KWREU": "{heli^}",
-  "HEP/A*T/SKWRO": "{hepato^}",
-  "HEPL/KWREU": "{hemi^}",
-  "HEPL/SKWREU": "{hemi^}",
-  "HEPT": "{hepato^}",
-  "HEPT/KWRO": "{hepato^}",
-  "HEPT/SKWRO": "{hepato^}",
-  "HERT": "{hetero^}",
-  "HET/RO": "{hetero^}",
-  "HET/ROE": "{hetero^}",
-  "HEUFT/RO": "{hystero^}",
-  "HEUFT/SKWRO": "{histo^}",
-  "HEUP/TPHO": "{hypno^}",
-  "HO*EUP": "{hypo^}",
-  "HO*EUPL": "{hemi^}",
-  "HO*EUPT": "{hepato^}",
-  "HO/HRO": "{holo^}",
-  "HO/PHO": "{homo^}",
-  "HOE/PHOE": "{homo^}",
-  "HOEPL/KWROE": "{homeo^}",
-  "HOEUP": "{hypo^}",
-  "HOEURD": "{hydro^}",
-  "HR*EUFPL/SKWRO": "{lympho^}",
-  "HR*EUP": "{lipo^}",
-  "HRA/PHEL/HRO": "{lamello^}",
-  "HRABGT/SKWRO": "{lacto^}",
-  "HRAEURPBG/KWRO": "{laryngo^}",
-  "HRAEURPBG/SKWRO": "{laryngo^}",
-  "HRAOUBG/SKWRO": "{leuko^}",
-  "HREBG/SKWRO": "{electro^}",
-  "HREBG/TRO": "{electro^}",
-  "HREPT/KWRO": "{lepto^}",
-  "HREPT/SKWRO": "{lepto^}",
-  "HRERBGT": "{electro^}",
-  "HRERBGT/SKWRO": "{electro^}",
-  "HRES/EPB": "{lessen^}",
-  "HREUP/SKWRO": "{lipo^}",
-  "HREUPBG/WO": "{linguo^}",
-  "HRO*BGS": "{hydroxy^}",
-  "HRO*EUFPL": "{lympho^}",
-  "HRO*EUFPL/SKWRO": "{lympho^}",
-  "HRO*EUP": "{lipo^}",
-  "HRUPL/PWO": "{lumbo^}",
-  "HRUPL/PWRO": "{lumbo^}",
-  "K*EUL": "{kilo^}",
-  "KA": "{ka^}",
-  "KA*LG/TPHAEU/KWROE": "{calcaneo^}",
-  "KA*LG/TPHO*EU": "{calcaneo^}",
-  "KAEUR/KWRO": "{karyo^}",
-  "KAEUR/KWROE": "{karyo^}",
-  "KAL/KAPB/KWRO": "{calcaneo^}",
-  "KAL/KAPB/KWROE": "{calcaneo^}",
-  "KAL/SEU": "{calci^}",
-  "KAOEPL": "{chemo^}",
-  "KAOEUL/SKWRO": "{chylo^}",
-  "KAOUPB/KWRO": "{cuneo^}",
-  "KAPS/HRO": "{capsulo^}",
-  "KARB/A/PHO*EUPB": "{carbamino^}",
-  "KARD/KWRO": "{cardio^}",
-  "KAUPB": "{con^}",
-  "KAURPBT": "{contra^}",
-  "KER/TO": "{kerato^}",
-  "KERT/KWRO": "{kerato^}",
-  "KERT/SKWRO": "{kerato^}",
-  "KET/KWRO": "{keto^}",
-  "KHAO*EPL": "{chemo^}",
-  "KHAOEPL": "{chemo^}",
-  "KHAOEPL/KWRO": "{chemo^}",
-  "KHAOEPL/SKWRO": "{chemo^}",
-  "KHAOEU/HRO": "{chylo^}",
-  "KHAOEUL/KWRO": "{chylo^}",
-  "KHAOEUL/SKWRO": "{chylo^}",
-  "KHAOEUPL/KWRO": "{chymo^}",
-  "KHAOEUPL/SKWRO": "{chymo^}",
-  "KHAOEUR/SKWRO": "{chiro^}",
-  "KHEPL/SKWRO": "{chemo^}",
-  "KHO*EUPL": "{chemo^}",
-  "KHOL/KREUFT": "{cholecyst^}",
-  "KHOPBD/RO": "{chondro^}",
-  "KHOR/KWROE": "{choreo^}",
-  "KHORPBD": "{chondro^}",
-  "KHR*EURT/RO": "{clarithro^}",
-  "KHRAEUD/KWRO": "{cleido^}",
-  "KHRAOED/KWRO": "{cleido^}",
-  "KHREFT/RO": "{cholestero^}",
-  "KHRO*R/SKWRO": "{chloro^}",
-  "KHROR/SKWRO": "{chloro^}",
-  "KOBG/SKWRO": "{cocco^}",
-  "KOEL/KWROE": "{choleo^}",
-  "KOEL/SKWRO": "{colo^}",
-  "KOEPL/A": "{comba^}",
-  "KOEURD": "{cardio^}",
-  "KOL/SKWRO": "{colo^}",
-  "KOPB/SKWRO": "{cono^}",
-  "KOPB/TKRO": "{chondro^}",
-  "KOPBD/RO": "{chondro^}",
-  "KORPB/KWRO": "{corneo^}",
-  "KORPB/SKWRE": "{coryne^}",
-  "KORPBT": "{contra^}",
-  "KORT/KO": "{cortico^}",
-  "KORT/KOE": "{cortico^}",
-  "KOUFRPBT": "{counter^}",
-  "KOURPBT": "{counter^}",
-  "KPAOE/TPHO": "{xeno^}",
-  "KPAOE/TPHOE": "{xeno^}",
-  "KPAOEPB/SKWRO": "{xeno^}",
-  "KPAOEUL/SKWRO": "{xylo^}",
-  "KPAOEUT/KWRO": "{excito^}",
-  "KPAOEUT/SKWRO": "{excito^}",
-  "KPWHROES": "{meso^}",
-  "KPWHROS/SO": "{glosso^}",
-  "KR*EF": "{cef^}",
-  "KRAEUPB/KWRO": "{cranio^}",
-  "KRAEUPB/KWROE": "{cranio^}",
-  "KRAEUPB/SKWRO": "{cranio^}",
-  "KRAO*EUT/KWRO": "{cyto^}",
-  "KRAO*EUT/SKWRO": "{cyto^}",
-  "KRAOEU/KWROE": "{cryo^}",
-  "KRAOEU/SKWRO": "{cryo^}",
-  "KRAOEUT/KWRO": "{cyto^}",
-  "KRAOEUT/SKWRO": "{cyto^}",
-  "KRAPB": "{cran^}",
-  "KREF": "{cef^}",
-  "KREUBG/KWRO": "{crico^}",
-  "KREUBG/SKWRO": "{crico^}",
-  "KREUFT/KWRO": "{cysto^}",
-  "KREUFT/SKWRO": "{cysto^}",
-  "KREUPT/SKWRO": "{crypto^}",
-  "KRO*EUT": "{cyto^}",
-  "KRO/PHO": "{chromo^}",
-  "KROEUT": "{cyto^}",
-  "KWAD/REU": "{quadri^}",
-  "KWARD": "{quadri^}",
-  "KWR*EU": "{i^}",
-  "KWRAO*E": "{e^}",
-  "KWRAT/RO": "{iatro^}",
-  "KWRES/TER": "{yester^}",
-  "KWREUBG/KWRO": "{syringo^}",
-  "KWROEPBD": "{endo^}",
-  "KWROUPB": "{down^}",
-  "O*EUBGS": "{extra^}",
-  "O*EUPBD": "{endo^}",
-  "O*EUPBG/SKWRO": "{onycho^}",
-  "O*EUR": "{uro^}",
-  "O*EURBGS": "{extra^}",
-  "O*EURT": "{erythro^}",
-  "O*EUT": "{auto^}",
-  "O*PBG": "{onco^}",
-  "O*PBG/HRO": "{onycho^}",
-  "O*PBG/SKWRO": "{onco^}",
-  "O*RT": "{ortho^}",
-  "O*RT/SKWRO": "{ortho^}",
-  "OBG/HRO": "{oculo^}",
-  "OBG/SKWRU": "{Ocu^}",
-  "OBGS/HRO": "{oxalo^}",
-  "OBGS/KWREU": "{oxy^}",
-  "OBGS/KWREU/TKO": "{oxido^}",
-  "OE/HREU/TKPWO": "{oligo^}",
-  "OE/TKOPBT/SKWRO": "{odonto^}",
-  "OEL/SRO": "{olivo^}",
-  "OEL/TKPWO": "{oligo^}",
-  "OEL/TKPWOE": "{oligo^}",
-  "OEP/KWROE": "{opio^}",
-  "OERBGS/SKWR-RBGS": "{oh^}",
-  "OET/KWRO": "{oto^}",
-  "OEUPBG/SKWRO": "{onycho^}",
-  "OEUR": "{uro^}",
-  "OFPL/HRO": "{omphalo^}",
-  "OFT/KWRO": "{osteo^}",
-  "OFT/KWROE": "{osteo^}",
-  "OFT/SKWRO": "{osteo^}",
-  "OP/THAL/PHO": "{ophthalmo^}",
-  "OPB/*EUBG": "{onycho^}",
-  "OPB/KWREU/KO": "{onycho^}",
-  "OPBG/SKWRO*": "{onco^}",
-  "OPL/TPHEU": "{omni^}",
-  "OR/SKWRO": "{oro^}",
-  "ORT/SKWRO": "{ortho^}",
-  "OS/PHO": "{osmo^}",
-  "OS/TAOE/KWROE": "{osteo^}",
-  "OT/KWRO": "{oto^}",
-  "OT/SKWRO": "{oto^}",
-  "P*ER": "{peri^}",
-  "PA*R": "{para^}",
-  "PA*T/SKWRO": "{patho^}",
-  "PABG/KWREU": "{pachy^}",
-  "PABG/SKWREU": "{pachy^}",
-  "PAEUL/KWRO": "{paleo^}",
-  "PAEUL/KWROE": "{paleo^}",
-  "PAL/TO": "{palato^}",
-  "PAOEU/HRO": "{pyelo^}",
-  "PAOEU/KWRO": "{pyo^}",
-  "PAOEUL/KWRO": "{pyelo^}",
-  "PAOEUL/SKPWRO": "{pyelo^}",
-  "PAOEUL/SKWRO": "{pyelo^}",
-  "PAOEUR/KWRO": "{pyro^}",
-  "PAOEUR/SKWRO": "{pyro^}",
-  "PAOUB/KWRO": "{pubo^}",
-  "PAOUB/SKWRO": "{pubo^}",
-  "PE/TKEUBG/HRO": "{pediculi^}",
-  "PEBG/TAOR/KWROE": "{bacterio^}",
-  "PEPT/SKWRO": "{pepto^}",
-  "PER/KWRO": "{perio^}",
-  "PER/OBGS": "{perox^}",
-  "PER/SKWREU": "{peri^}",
-  "PET/RO": "{petro^}",
-  "PEUBG/SKWRO": "{pico^}",
-  "PEUBGT/SKWRO": "{picto^}",
-  "PEUL/KWRO": "{pilo^}",
-  "PEUL/SKWRO": "{pilo^}",
-  "PH*EL/TPHO": "{melano^}",
-  "PH*ELT/HRO": "{metallo^}",
-  "PH*ELT/SKWRO": "{metallo^}",
-  "PH*ERT": "{metro^}",
-  "PH*ET/*L": "{methyl^}",
-  "PH*EUL": "{milli^}",
-  "PH*UFL/HRO": "{musculo^}",
-  "PH-BG": "{Mc^}",
-  "PHABG/HRO": "{maculo^}",
-  "PHABGS/HRO": "{maxillo^}",
-  "PHAERBG": "{macro^}",
-  "PHAL": "{mal^}",
-  "PHAO*UPB": "{immuno^}",
-  "PHAOED/KWRO": "{medio^}",
-  "PHAOES/KWRO": "{mesio^}",
-  "PHAOEU/HRO": "{myelo^}",
-  "PHAOEU/KROE": "{micro^}",
-  "PHAOEU/KWRO": "{myo^}",
-  "PHAOEU/SKWRO": "{myo^}",
-  "PHAOEUBG/KWRO": "{myco^}",
-  "PHAOEUBG/SKPWRO": "{myco^}",
-  "PHAOEUBG/SKWRO": "{myco^}",
-  "PHAOEUL/KWRO": "{myelo^}",
-  "PHAOEUL/SKWRO": "{myelo^}",
-  "PHAOEURBG": "{micro^}",
-  "PHAOU/KOE": "{muco^}",
-  "PHAOUBG": "{muco^}",
-  "PHAOUBG/KO": "{muco^}",
-  "PHAOUBG/SKWRO": "{muco^}",
-  "PHAPL/HRO": "{mammillo^}",
-  "PHE/TAL/HRO": "{metallo^}",
-  "PHE/TPHEUPBG": "{meningo^}",
-  "PHEBG/TPHO": "{mechano^}",
-  "PHEG/HRO": "{megalo^}",
-  "PHEG/KWRA": "{mega^}",
-  "PHEG/TKPWA": "{mega^}",
-  "PHEL/TPHO": "{melano^}",
-  "PHEPB/EUPB/SKWRO": "{meningo^}",
-  "PHEPB/EUS/KO": "{menisco^}",
-  "PHEPBG/SKWRO": "{meningo^}",
-  "PHES/KWRO": "{meso^}",
-  "PHES/SKWRO": "{meso^}",
-  "PHET/HRO": "{metallo^}",
-  "PHET/KWRA": "{meta^}",
-  "PHEUD": "{mid^}",
-  "PHEUPB/SKWRO": "{mino^}",
-  "PHEURPBG": "{myringo^}",
-  "PHEUZ": "{mis^}",
-  "PHO*EU": "{myo^}",
-  "PHO*EUPB/KRAOEUT/KWRO": "{immunocyto^}",
-  "PHO*EUT": "{meta^}",
-  "PHO*FR/KWRO": "{morpho^}",
-  "PHO*PB": "{mono^}",
-  "PHO/TO": "{moto^}",
-  "PHOPB/OE": "{mono^}",
-  "PHOPB/SKWRO": "{mono^}",
-  "PHRAEU/KWRO": "{pleo^}",
-  "PHRAOE/KWRO": "{pleio^}",
-  "PHRAS": "{plas^}",
-  "PHRAT/KWREU": "{platy^}",
-  "PHRE/KWRO": "{pleo^}",
-  "PHRE/SKWRO": "{pleo^}",
-  "PHRO*EUPB": "{immuno^}",
-  "PHRUR/SKWREU": "{pluri^}",
-  "PHUFBG/HRO": "{musculo^}",
-  "PO*EFT/RO": "{postero^}",
-  "PO*EUBG": "{pico^}",
-  "PO*EUT": "{patho^}",
-  "PO*L": "{poly^}",
-  "POEFT/RO": "{postero^}",
-  "POFT/RO": "{postero^}",
-  "POL/KWREU": "{poly^}",
-  "POPBT/SKWRO": "{ponto^}",
-  "PRAOEUT/SKWRO": "{parieto^}",
-  "PRO": "{pro^}",
-  "PROBGT/SKWRO": "{procto^}",
-  "PROEP/RO": "{proprio^}",
-  "PROERP/RO": "{proprio^}",
-  "PROET/KWRO": "{proteo^}",
-  "PROET/KWROE": "{proteo^}",
-  "PW*EU": "{bi^}",
-  "PWA*EBG": "{back^}",
-  "PWABG/TAOER/KWRO": "{bacterio^}",
-  "PWAEBG": "{back^}",
-  "PWAEU/TKPWA": "{bayga^}",
-  "PWAEUR/SKWRO": "{baro^}",
-  "PWAEUS/SKWRO": "{baso^}",
-  "PWAO*E": "{bio^}",
-  "PWAOEBGT/RO": "{bacterio^}",
-  "PWAOERBGT/KWRO": "{bacterio^}",
-  "PWAOERBGT/SKWRO": "{bacterio^}",
-  "PWAOEU/KWROE": "{bio^}",
-  "PWAOUBG/KWRO": "{bucco^}",
-  "PWAOUT/RO": "{butyro^}",
-  "PWAR/SKWRO": "{baro^}",
-  "PWE": "{be^}",
-  "PWEPB/SO*": "{benzo^}",
-  "PWHRAFT/SKWRO": "{blasto^}",
-  "PWHREF/RO": "{blepharo^}",
-  "PWHREFR/SKWRO": "{blepharo^}",
-  "PWO*E": "{bio^}",
-  "PWO*EU": "{bio^}",
-  "PWRA*EUD/KWREU": "{brady^}",
-  "PWRA*PBG/KWRO": "{branchio^}",
-  "PWRABG/KWREU": "{brachy^}",
-  "PWRAD/KWREU": "{brady^}",
-  "PWRAEUBG/KWRO": "{brachio^}",
-  "PWRAPBG/KWRO": "{branchio^}",
-  "PWRO*PBG": "{broncho^}",
-  "PWRO*PBG/KWRO": "{bronchio^}",
-  "PWRO*PBG/SKWRO": "{broncho^}",
-  "PWROPB/KO": "{broncho^}",
-  "PWROPBG/KWRO": "{broncho^}",
-  "PWROPBG/SKWRO": "{broncho^}",
-  "PWUBG/KO": "{bucco^}",
-  "PWUBG/KWRO": "{bucco^}",
-  "PWUBG/SKWRO": "{bucco^}",
-  "PWUBL/KWRO": "{bulbo^}",
-  "PWUBL/SKWRO": "{bulbo^}",
-  "PWUT/OBG/SKWRU": "{but Ocu^}",
-  "R*ERT": "{erythro^}",
-  "RAOEPB": "{reen^}",
-  "RAOEU/PWO": "{ribo^}",
-  "RAOEUB/SKWRO": "{ribo^}",
-  "RAOEUB/TPHAOU/KHRAEU/SKWRO": "{ribonucleo^}",
-  "RAOEUPB/SKWRO": "{rhino^}",
-  "RAOUB/RO": "{rubro^}",
-  "RAOUPL/TO": "{rheumato^}",
-  "RAOUPLT": "{rheumato^}",
-  "RAOUPLT/SKWRO": "{rheumato^}",
-  "RE": "{re^}",
-  "RE/TEUBG/HRO": "{reticulo^}",
-  "REBGT/KWRO": "{recto^}",
-  "REBGT/SKWRO": "{recto^}",
-  "RERT": "{retro^}",
-  "RET/TPHO": "{retino^}",
-  "REUB/SKWRO": "{ribo^}",
-  "REUPB/SKWRO": "{rhino^}",
-  "RO/PWO": "{robo^}",
-  "ROEPBT": "{entero^}",
-  "S*EUPB": "{syn^}",
-  "S*EUPL/A*T/SKWRO": "{sympatho^}",
-  "S*EUPL/THO": "{sympatho^}",
-  "S*EUPLT/SKWRO": "{sympatho^}",
-  "SABG/RO": "{sacro^}",
-  "SAL/PEUPB/TKPWO": "{salpingo^}",
-  "SAO*EU/TKPWO": "{zygo^}",
-  "SAO*EUG": "{zyg^}",
-  "SAO*EUG/KWRO/PHAT/K-L": "{zygomatico^}",
-  "SAO*EUG/PHAT/KO": "{zygomatico^}",
-  "SAO*UD": "{pseudo^}",
-  "SAOEUB/*ER": "{cyber^}",
-  "SAOEUBG/HRO": "{cyclo^}",
-  "SAOEUD/RO": "{sidero^}",
-  "SAOEUG/KWRO": "{zygo^}",
-  "SAOEUG/SKWRO": "{zygo^}",
-  "SAOEUG/SKWRO/PHAT/K-L": "{zygomatico^}",
-  "SAOEULG": "{cyclo^}",
-  "SAOEUPB/KWRO": "{cyano^}",
-  "SAOEUPB/SKWRO": "{cyano^}",
-  "SAOEURB": "{cyber^}",
-  "SAOEUS/SKWRO": "{iso^}",
-  "SAOEUT/O*": "{cyto^}",
-  "SAOEUT/SKWRO": "{cyto^}",
-  "SAOUD/KWRO": "{pseudo^}",
-  "SAOUD/SKWRO": "{pseudo^}",
-  "SAOUD/TKOE": "{pseudo^}",
-  "SAOUP": "{super^}",
-  "SAOUP/RA": "{supra^}",
-  "SAOUP/RO": "{supero^}",
-  "SAR/KO": "{sarco^}",
-  "SARBG": "{sarco^}",
-  "SARBG/SKWRO": "{sarco^}",
-  "SAUB": "{sub^}",
-  "SE/RAOE/PWROE": "{cerebro^}",
-  "SEF/HRO": "{cephalo^}",
-  "SEFL/SKWRO": "{cephalo^}",
-  "SEFRBG/SKWRO": "{cervico^}",
-  "SEPB/TRO": "{centro^}",
-  "SEPLT/SKWRO": "{cemento^}",
-  "SER/HRO": "{cerulo^}",
-  "SER/PWEL/HRO": "{cerebello^}",
-  "SER/PWEL/SKWRO": "{cerebello^}",
-  "SER/SKWRO": "{sero^}",
-  "SEUPB/KROE": "{synchro^}",
-  "SEUPL/PA*T/SKWRO": "{sympatho^}",
-  "SEUPL/PA/THO": "{sympatho^}",
-  "SEUPL/PO*EUT": "{sympatho^}",
-  "SEURBG/UPL": "{circum^}",
-  "SEUT/RO": "{citro^}",
-  "SHO*E": "{show^}",
-  "SKHRER/SKWRO": "{sclero^}",
-  "SKRAOEB/RO": "{cerebro^}",
-  "SKWAEUPL/SKWRO": "{squamo^}",
-  "SKWRAOE/KWROE": "{geo^}",
-  "SKWRAOE/OE": "{geo^}",
-  "SKWRAOEPB/KWRO": "{genio^}",
-  "SKWRAOU/TKAEU/KWROE": "{Judeo^}",
-  "SKWRAOUP/*ER": "{juper^}",
-  "SKWREPBT/KWRO": "{genito^}",
-  "SKWREPBT/SKWRO": "{genito^}",
-  "SKWRO*EU": "{geo^}",
-  "SKWROEPBD": "{endo^}",
-  "SKWRUG/HRO": "{jugulo^}",
-  "SKWRUGT": "{juxta^}",
-  "SKWRUGT/HRO": "{juxtallo^}",
-  "SKWRUGT/KWRA": "{juxta^}",
-  "SO*EUBG": "{psycho^}",
-  "SO/PHA/TO": "{somato^}",
-  "SOE/PHAT/SKWRO": "{somato^}",
-  "SOEPL/TO": "{somato^}",
-  "SOEPLT": "{somato^}",
-  "SOEPLT/KWRO": "{somato^}",
-  "SOEPLT/SKWRO": "{somato^}",
-  "SOEPLTS/SKWRO": "{somato^}",
-  "SOERB/KWROE": "{socio^}",
-  "SOES/KWROE": "{socio^}",
-  "SPAOEUPB/KWRO": "{spino^}",
-  "SPAOEUPB/SKWRO": "{spino^}",
-  "SPAS/PHO": "{spasmo^}",
-  "SPERBGT": "{spectro^}",
-  "SPEUPB/SKWRO": "{spino^}",
-  "SPHEU": "{semi^}",
-  "SPOPBD/HRO": "{spondylo^}",
-  "SPOPBG/KWROE": "{spongio^}",
-  "SPOR/SKWRO": "{sporo^}",
-  "SPWRA": "{intra^}",
-  "SRAEUG/SKWRO": "{vago^}",
-  "SRAEUS/KWRO": "{vaso^}",
-  "SRAEUS/SKWRO": "{vaso^}",
-  "SRAEUT/SKWRO": "{vaso^}",
-  "SRAFBG/HRO": "{vasculo^}",
-  "SRAFL/SKWRO": "{valvulo^}",
-  "SRAOEB/ROE": "{cerebro^}",
-  "SRAOEPB/KWRO": "{veno^}",
-  "SRAOEPB/SKWRO": "{veno^}",
-  "SRAOEUB/RA": "{vibra^}",
-  "SRAS/SKWRO": "{vaso^}",
-  "SREFT/PWHRO": "{vestibulo^}",
-  "SREL/SRO": "{vulvo^}",
-  "SREPB/KWRO": "{veno^}",
-  "SREPB/SKWRO": "{veno^}",
-  "SREPB/TRO": "{ventro^}",
-  "SREPBT/RO": "{ventro^}",
-  "SRES/KO": "{vesico^}",
-  "SREUFBG/SKWRO": "{visco^}",
-  "SREUPBG/KWRO": "{syringo^}",
-  "SREURT/KWRO": "{vitreo^}",
-  "SREURT/RO": "{vitreo^}",
-  "SREURT/SKWRO": "{vitreo^}",
-  "SREUS/KWRO": "{visio^}",
-  "SREUS/RO": "{viscero^}",
-  "SREUS/SKWRO": "{visuo^}",
-  "SRO*EUS": "{vaso^}",
-  "STAF/HRO": "{staphylo^}",
-  "STAOET/SKWRO": "{steato^}",
-  "STAOEU/HRO": "{stylo^}",
-  "STAOEUL/KWRO": "{stylo^}",
-  "STAOEUL/SKWRO": "{stylo^}",
-  "STEPT": "{strepto^}",
-  "STER/KWRO": "{stereo^}",
-  "STERBG/KWRO": "{sterco^}",
-  "STERPB/SKWRO": "{sterno^}",
-  "STKPWAOEU/TKPWO": "{zygo^}",
-  "STKPWAOEUG/SKWRO": "{zygo^}",
-  "STO*EUL": "{stylo^}",
-  "STOEPL/TO": "{stomato^}",
-  "STPAOEPB/KWRO": "{spheno^}",
-  "STPAOEPB/SKWRO": "{spheno^}",
-  "STPAOER/SKWRO": "{sphero^}",
-  "STPEUPBG/KWRO": "{sphingo^}",
-  "STPEUPBG/SKWRO": "{sphingo^}",
-  "STPHAPT/SKWRO": "{synapto^}",
-  "STPO*EUPB": "{spheno^}",
-  "STREPT/SKWRO": "{strepto^}",
-  "STREPT/SKWRO*": "{strepto^}",
-  "SUP/RA": "{supra^}",
-  "T*EL": "{tele^}",
-  "T*EPL/RO": "{temporo^}",
-  "T*ERT": "{tetra^}",
-  "TA*BG/KWREU": "{tachy^}",
-  "TA*BG/SKWREU": "{tachy^}",
-  "TABG/SKWREU": "{tachy^}",
-  "TAER/TKPWO": "{pterygo^}",
-  "TAL/KAPB/KAPB/KWRAL": "{talocancaneal^}",
-  "TAL/KWRO": "{talo^}",
-  "TAL/SKWRO": "{talo^}",
-  "TAOUB/HRO": "{tubulo^}",
-  "TAOUB/RO": "{tubero^}",
-  "TAUR": "{tor^}",
-  "TER/EUPBLG/KWRO": "{pterygo^}",
-  "TER/TKPWO": "{pterygo^}",
-  "TER/TO": "{terato^}",
-  "TERPBLG/KWRO": "{pterygo^}",
-  "TERT/KWRO": "{terato^}",
-  "TERT/SKWRO": "{terato^}",
-  "TET/RA": "{tetra^}",
-  "TET/TRA": "{tetra^}",
-  "TEUB/KWRO": "{tibio^}",
-  "THAL/PHO": "{thalamo^}",
-  "THAOEU/RO": "{thyro^}",
-  "THAOEUR/SKWRO": "{thyro^}",
-  "THERPL/KWRA": "{therma^}",
-  "THO*EU": "{theo^}",
-  "THOEU": "{thio^}",
-  "THOR/KO": "{thoraco^}",
-  "THRO*PL/SKWRO": "{thrombo^}",
-  "THROPL": "{thrombo^}",
-  "THROPL/PWO": "{thrombo^}",
-  "THROPL/SKWRO": "{thrombo^}",
-  "TK*EPL": "{demi^}",
-  "TK*ERPL/TO": "{dermato^}",
-  "TKAERBG/KWRO": "{dacryo^}",
-  "TKAOEU/KWRA": "{dia^}",
-  "TKARBG/KWRO": "{dacryo^}",
-  "TKARBG/KWROE": "{dacryo^}",
-  "TKARBG/RO": "{dacryo^}",
-  "TKE": "{de^}",
-  "TKEBGS/TRO": "{dextro^}",
-  "TKEG/SKWRA": "{deca^}",
-  "TKEPB/TKO": "{dendro^}",
-  "TKEPB/TKRO": "{dendro^}",
-  "TKERPL/AT/SKWRO": "{dermato^}",
-  "TKERPL/SKWRO": "{dermo^}",
-  "TKERPLT/KWRO": "{dermato^}",
-  "TKERPLT/SKWRO": "{dermato^}",
-  "TKEUBGS": "{diction^}",
-  "TKEUFT/KWRO": "{disto^}",
-  "TKEUFT/SKWRO": "{disto^}",
-  "TKEUS": "{dis^}",
-  "TKEUS/TKAOEU/TKO": "{dysdiado^}",
-  "TKOR/SO": "{dorso^}",
-  "TKORS/KWRO": "{dorso^}",
-  "TKORS/SKWREU": "{dorsi^}",
-  "TKORS/SKWRO": "{dorso^}",
-  "TKPWA/HRABGT": "{galacto^}",
-  "TKPWAFT/RO": "{gastro^}",
-  "TKPWAPL/AOET/SKWRO": "{gameto^}",
-  "TKPWAS/TRO": "{gastro^}",
-  "TKPWEUG/KWRA": "{giga^}",
-  "TKPWHR*/HR*/*U/KR*/O*": "{gluco^}",
-  "TKPWHRABG/TO": "{galacto^}",
-  "TKPWHRABGT": "{galacto^}",
-  "TKPWHRABGT/SKWRO": "{galacto^}",
-  "TKPWHRAOEU/KO": "{glyco^}",
-  "TKPWHRAOEUBG": "{glyco^}",
-  "TKPWHRAOEUBG/KO": "{glyco^}",
-  "TKPWHRAOEUBG/SKWRO": "{glyco^}",
-  "TKPWHRAOU/KOE": "{gluco^}",
-  "TKPWHRAOUBG": "{gluco^}",
-  "TKPWHRAOUBG/SKWRO": "{gluco^}",
-  "TKPWHRAOUBG/TPHO": "{glucono^}",
-  "TKPWHRO/PHER/HRO": "{glomerulo^}",
-  "TKPWHROPL/*ER/HRO": "{glomerulo^}",
-  "TKPWHROPL/HRO": "{glomerulo^}",
-  "TKPWHROPL/RU": "{glomerulo^}",
-  "TKPWHROS/SKWRO": "{glosso^}",
-  "TKPWO*EU": "{geo^}",
-  "TKPWOLD": "{gold^}",
-  "TKPWRAPB/HRO": "{granulo^}",
-  "TKWOD/TPHO": "{duodeno^}",
-  "TO*P/SKWRO": "{topo^}",
-  "TO/A*EUPBLT": "{to anti^}",
-  "TO/PO": "{topo^}",
-  "TOBGS/KO": "{toxico^}",
-  "TPAEUG/KWRO": "{phago^}",
-  "TPAEUG/SKWRO": "{phago^}",
-  "TPAEUPBLG/SKWRO": "{phago^}",
-  "TPAEURB/KWRO": "{facio^}",
-  "TPAEURPBG/KWRO": "{pharyngo^}",
-  "TPAEURPBG/SKWRO": "{pharyngo^}",
-  "TPAEUS/KWRO": "{facio^}",
-  "TPAL/SKWRO": "{phallo^}",
-  "TPAOEPB/*L": "{phenyl^}",
-  "TPAOEPBL": "{phenyl^}",
-  "TPAOET/KWRO": "{feto^}",
-  "TPAOET/SKWRO": "{feto^}",
-  "TPAOEUB/RO": "{fibro^}",
-  "TPAOEUB/ROE": "{fibro^}",
-  "TPAOEURB/RO": "{fibro^}",
-  "TPAOEUT/KWRO": "{phyto^}",
-  "TPAOEUT/SKWRO": "{phyto^}",
-  "TPAOUS/KWRO": "{fuso^}",
-  "TPARBG/KO": "{pharmaco^}",
-  "TPARPL/KO": "{pharmaco^}",
-  "TPARPL/KOE": "{pharmaco^}",
-  "TPAUR": "{for^}",
-  "TPEPLT/SKWRO": "{femto^}",
-  "TPER/ROE": "{ferro^}",
-  "TPER/SKWRO": "{fero^}",
-  "TPEUB/HRO": "{fibulo^}",
-  "TPEUS/KWRO": "{physio^}",
-  "TPEUS/KWROE": "{physio^}",
-  "TPEUS/SKWRO": "{physio^}",
-  "TPH*UR": "{neuro^}",
-  "TPHAEUS/SKWRO": "{naso^}",
-  "TPHAO*UR": "{neuro^}",
-  "TPHAOEUG/RO": "{nigro^}",
-  "TPHAOEURT/SKWRO": "{nitro^}",
-  "TPHAOEUT/RO": "{nitro^}",
-  "TPHAOU/ROE": "{neuro^}",
-  "TPHAOUBG/HRO": "{nucleo^}",
-  "TPHAOUBG/HROE": "{nucleo^}",
-  "TPHAOUBG/SKWRO": "{nucho^}",
-  "TPHAOUPL": "{pneumo^}",
-  "TPHAOUPL/PHOE": "{pneumo^}",
-  "TPHAOUPL/SKWRO": "{pneumo^}",
-  "TPHAOUR/SKWRO": "{neuro^}",
-  "TPHAOURT/SKWRO": "{neutro^}",
-  "TPHAPB/KWRO": "{nano^}",
-  "TPHAPB/SKWRO": "{nano^}",
-  "TPHAUPB": "{non^}",
-  "TPHEBG/RO": "{necro^}",
-  "TPHEF/RO": "{nephro^}",
-  "TPHOD/HRO": "{nodulo^}",
-  "TPHORPL/SKWRO": "{normo^}",
-  "TPHRAO*UR": "{fluoro^}",
-  "TPHRAOUBG/HRO": "{nucleo^}",
-  "TPHRAOUD/RO": "{fludro^}",
-  "TPHRAOUR": "{fluoro^}",
-  "TPHROR/SKWRO": "{fluoro^}",
-  "TPHUR/SKWRO": "{neuro^}",
-  "TPO/TO": "{photo^}",
-  "TPO/TPHO": "{phono^}",
-  "TPOER": "{fore^}",
-  "TPOEURB": "{fibro^}",
-  "TPORPL/*L": "{formyl^}",
-  "TPOS/TO": "{phospho^}",
-  "TPOS/TPA": "{phospha^}",
-  "TPOS/TPO": "{phospho^}",
-  "TPOS/TPOE": "{phospho^}",
-  "TPRAEUPB": "{infra^}",
-  "TPRAOEPB/KWRO": "{phreno^}",
-  "TPRAOUBGT": "{fructo^}",
-  "TPRUBG/TO": "{fructo^}",
-  "TR*EU": "{tri^}",
-  "TR*EU/AOEU/TKO": "{triiodo^}",
-  "TR*EU/SKWREPL/TPHO": "{trigemino^}",
-  "TR*EU/SKWRO": "{trio^}",
-  "TR*EUP/TPHO": "{trypano^}",
-  "TRAEUBG/KWRO": "{tracheo^}",
-  "TRAO*EU": "{tri^}",
-  "TRAPBS/SRERS/SKWRO": "{transverso^}",
-  "TRAPBZ": "{trans^}",
-  "TRERT": "{tetra^}",
-  "TRO/PO": "{tropo^}",
-  "TRO/TO": "{tropo^}",
-  "TROEF/SKWRO": "{tropho^}",
-  "TROF/SKWRO": "{tropho^}",
-  "TWOUPB": "{200^}",
-  "UL/TRA": "{ultra^}",
-  "UPB": "{un^}",
-  "UPBD": "{under^}",
-  "UPL/PHAOUPB/TPHO": "{immuno^}",
-  "UPL/PWEUL/KO": "{umbilico^}",
-  "UR/KWRO": "{uro^}",
-  "UR/SKWRO": "{uro^}",
-  "URLT": "{ultra^}",
-  "WA*UR": "{water^}",
-  "A*UF": "{off-^}",
-  "AEUPB/SKWREU": "{ani-^}",
-  "AFR/ROE": "{Afro-^}",
-  "AOE/KOE": "{eco-^}",
-  "AOEBG/WEU/HRAT/RAL": "{equilateral-^}",
-  "EUPBD/SKWRO": "{Indo-^}",
-  "HA*EPBD": "{hand-^}",
-  "HAEPBD": "{handi-^}",
-  "HAOEU/PWERPB/TPHOE": "{Hiberno-^}",
-  "HER/SEF": "{herself-^}",
-  "HREUS": "{Lys-^}",
-  "KAU": "{co^}",
-  "KOE": "{co-^}",
-  "KWAS/KWREU": "{quasi-^}",
-  "KWAS/SKWREU": "{quasi-^}",
-  "KWAT/SKWREU": "{quasi-^}",
-  "KWR*EUPB": "{in-^}",
-  "KWRO*PB": "{on-^}",
-  "KWRO*UT": "{out-^}",
-  "PH*UFP": "{much-^}",
-  "PHAO*EURBG": "{micro-^}",
-  "PHAOEU/SEF": "{myself-^}",
-  "PHAUPB": "{non-^}",
-  "PRE": "{pre^}",
-  "PR*E": "{pre-^}",
-  "PRO/TO": "{proto-^}",
-  "PROE/TOE": "{proto-^}",
-  "R*E": "{re-^}",
-  "S*EPL": "{semi-^}",
-  "SA*UB": "{sub-^}",
-  "SAO*URP": "{super-^}",
-  "SEF": "{self-^}",
-  "SKWR*EL": "{well-^}",
-  "TKPWHRAOEU": "{gly-^}",
-  "TKPWRA*ET": "{great-^}",
-  "TKPWREBG/SKWRO": "{Greco-^}",
-  "TPHA*UPB": "{non-^}",
-  "TPHAOE/KWROE": "{neo-^}",
-  "TPHO*": "{no-^}",
-  "W*EL": "{well-^}",
-  "A*D/TPHO": "{adeno^}",
-  "A*UT/SKWRO": "{auto^}",
-  "A/SAO*ET/*L": "{acetyl^}",
-  "A/SAOE/TO": "{aceto^}",
-  "A/SAOES/*L": "{acetyl^}",
-  "A/TKAOE/TPHO": "{adeno^}",
-  "A/TKAOE/TPHOE": "{adeno^}",
-  "A/TKAOEPB/KWRO": "{adeno^}",
-  "A/TKRAOE/TPHO": "{adreno^}",
-  "A/TKRAOE/TPHROE": "{adreno^}",
-  "A/TKRAOEPB": "{adreno^}",
-  "A/TKRO*EUPB": "{adreno^}",
-  "AER/SKWRO": "{aero^}",
-  "AEUR/SKWRO": "{aero^}",
-  "KWRA*D": "{ad^}",
-  "RAOE": "{re^}",
-  "TPWOUPB": "{200^}",
-  // ["AO*UT/", "out"],
-  // ["AOPB/", "on"],
-  // ["APB/TAOEU/", "anti"],
-  // ["APB/TEU/", "anti"],
-  // ["KWROUT/", "out"],
-  // ["KWRUP/", "up"],
-  // ["KWRUPBD/", "under"],
-  // ["SAOUPS/", "super"],
-  // ["SPR/", "super"],
-  // ["SPWR/", "inter"],
-  // ["TKEUZ/", "dis"],
-  // ["UPBDZ/", "under"],
-  // ["KRO/", "co-"]
-};
-const testAussieDict = {
-  "AOEUZ/A*U": "{^ise}",
-  "AO*EUFD/A*U": "{^ised}",
-  "AO*EUFG/A*U": "{^ising}",
-  // "*LG": "{^ling}",
-  // "*LD": "{^led}",
-  "*LD/A*U": "{^led}",
-  "*LG/A*U": "{^ling}",
-  // "*EG": "{^eing}",
-  "*EG/A*U": "{^eing}",
-  "KWROR/A*U": "{^iour}",
-  "KWRO*R/A*U": "{^iour}",
-  "O*UR/A*U": "{^our}",
-}
-const testSuffixesDict = {
-  "*D": "{^'d}",
-  "AES": "{^'s}",
-  "*EPLT": "{^ement}",
-  "*BG": "{^k}",
-  "*EFBG": "{^esque}",
-  "*EPB": "{^en}",
-  "*EPB/TPELD": "{^enfeld}",
-  "*ER": "{^er}",
-  "*ERS": "{^ers}",
-  "*ES": "{^ess}",
-  "*ET": "{^eth}",
-  "*EUFPL": "{^ism}",
-  "*EUFT": "{^ivity}",
-  "*EUGS": "{^ician}",
-  "*EUL/EUBG": "{^ilic}",
-  "*EUPB": "{^in}",
-  "*EUPB/*EUPB": "{^inin}",
-  "*EUPBS": "{^iness}",
-  "*EUPBZ": "{^ins}",
-  "*EUS": "{^ice}",
-  "*EZ": "{^ez}",
-  "*FP": "{^ch}",
-  "*FPL": "{^sm}",
-  "*L": "{^le}",
-  "*P": "{^p}",
-  "*PB": "{^n}",
-  "*PBD": "{^nd}",
-  "*PBLG": "{^ge}",
-  "*PLT": "{^ment}",
-  "*RB": "{^sh}",
-  "*RDZ": "{^rds}",
-  "*T": "{^th}",
-  "*UPL": "{^um}",
-  "-BL": "{^able}",
-  "-BLT": "{^ability}",
-  "-D": "{^ed}",
-  "-FL": "{^ful}",
-  "-FPD": "{^ed}{.}",
-  "-FPS": "{^s}{.}",
-  "-G": "{^ing}",
-  "-LS": "{^less}",
-  "-PBS": "{^ness}",
-  "-PLTS": "{^ments}",
-  "-S": "{^s}",
-  "A*EL": "{^ally}",
-  "A*EUBG": "{^ache}",
-  "A*L": "{^al}",
-  "A*L/EUFPL": "{^alism}",
-  "A*L/EUS": "{^alis}",
-  "A*PB": "{^an}",
-  "A*PB/SKWREU": "{^ani}",
-  "A*PBL": "{^allian}",
-  "A*R": "{^ar}",
-  "A*R/A*ET": "{^arate}",
-  "A*RPL": "{^arum}",
-  "A*RS": "{^ars}",
-  "A*T": "{^ate}",
-  "A*T/-D": "{^ated}",
-  "A*T/EUBG": "{^atic}",
-  "A*T/EUF": "{^ative}",
-  "A*T/K-L": "{^atical}",
-  "A*TD": "{^ated}",
-  "A/HOL/EUBG": "{^aholic}",
-  "A/TA": "{^ata}",
-  "A/TOR/KWRUPL": "{^atorium}",
-  "AER": "{^ary}",
-  "AEUGS": "{^ation}",
-  "AEURBS": "{^aceous}",
-  "AEURL": "{^arial}",
-  "AEURPB": "{^arian}",
-  "AEURPL": "{^arium}",
-  "AEURT": "{^arity}",
-  "AEUT/-D": "{^ated}",
-  "AEUT/-G": "{^ating}",
-  "AEUT/O*R": "{^ator}",
-  "AEUZ": "{^ase}",
-  "ALG/KWRA": "{^algia}",
-  "ALT": "{^ality}",
-  "ALT/EUBG": "{^atic}",
-  "AO*E": "{^ie}",
-  "AO*EL": "{^elle}",
-  "AO*EPB": "{^ean}",
-  "AO*ER": "{^eer}",
-  "AO*ET": "{^ette}",
-  "AO*EUFD": "{^ized}",
-  "AO*EUFG": "{^izing}",
-  "AO*EUTS": "{^itis}",
-  "AO*EZ": "{^ese}",
-  "AO*UL/A*T/-D": "{^ulated}",
-  "AO*UR/KWRA": "{^uria}",
-  "AOE/TOE": "{^ito}",
-  "AOEPL/EUBG": "{^emic}",
-  "AOEPL/KWRA": "{^emia}",
-  "AOEPL/KWRARBGS": "{^emia}{,}",
-  "AOES/EUS": "{^esis}",
-  "AOES/KWRA": "{^esia}",
-  "AOEU/KWRA/SEUS": "{^iasis}",
-  "AOEU/SEUS": "{^iasis}",
-  "AOEUBL": "{^izable}",
-  "AOEUPB": "{^ine}",
-  "AOEUT": "{^ite}",
-  "AOEUT/EUS": "{^itis}",
-  "AOEUTS": "{^itis}",
-  "AOEUZ": "{^ize}",
-  "AOEZ": "{^ese}",
-  "AOF/ROPB": "{^oophoron}",
-  "AOUD/AEUGS": "{^udation}",
-  "AOUL": "{^ule}",
-  "AOUL/A*T": "{^ulate}",
-  "AOULS": "{^ules}",
-  "AOUPBLG": "{^uge}",
-  "AOUR": "{^ure}",
-  "APBLG": "{^age}",
-  "APBS": "{^ance}",
-  "APBT": "{^ant}",
-  "AR/HREU": "{^arily}",
-  "AS/TAFP": "{^s} attach", // FIXME: might not match affixList regex
-  "AT/AOEUZ/-D": "{^atized}",
-  "AT/EUBG": "{^atic}",
-  "AT/KHREU": "{^atically}",
-  "AZ/PAPL": "{^orazepam}",
-  "EBGT/PHEU": "{^ectomy}",
-  "EF/RAOEPB": "{^ephrine}",
-  "EF/REUPB": "{^ephrine}",
-  "EFT": "{^est}",
-  "EL/HRA": "{^ella}",
-  "EPBL": "{^ential}",
-  "EPBS": "{^ence}",
-  "EPBT": "{^ent}",
-  "EPBT/-D": "{^ented}",
-  "EPBZ/PAOEUPB": "{^enzepine}",
-  "ERPBLG/EUBG": "{^ergic}",
-  "ERS": "{^ers}",
-  "ES/EPBT": "{^escent}",
-  "ES/KEU": "{^eski}",
-  "ET": "{^et}",
-  "EUBG": "{^ic}",
-  "EUBG/HRAT": "{^iculate}",
-  "EUBGS": "{^ix}",
-  "EUBL": "{^ible}",
-  "EUBLT": "{^ibility}",
-  "EUD/AEUZ": "{^idase}",
-  "EUF": "{^ive}",
-  "EUFBG": "{^iffic}",
-  "EUFL": "{^ively}",
-  "EUFPBS": "{^iveness}",
-  "EUFPL": "{^ism}",
-  "EUFT": "{^ist}",
-  "EUGS": "{^ition}",
-  "EULT": "{^ility}",
-  "EUPB/*EU": "{^ini}",
-  "EUPB/KWRAPB": "{^inian}",
-  "EUPB/SKWREU": "{^ini}",
-  "EUPBG/*ER": "{^inger}",
-  "EUPL/TREU": "{^imetry}",
-  "EURB": "{^ish}",
-  "EURBS": "{^itious}",
-  "EUS": "{^is}",
-  "EUS/TEU": "{^icity}",
-  "EUS/TEUS": "{^icities}",
-  "EUT/KWRA": "{^ita}",
-  "EUT/KWREU": "{^ity}",
-  "H-FS": "{^s?}",
-  "HA*PBD": "{^hand}",
-  "HA*PL": "{^ham}",
-  "HAO*D": "{^hood}",
-  "HAOED/RAL": "{^hedral}",
-  "HO*ELD": "{^hold}",
-  "HO*ERLD": "{^holder}",
-  "HO*LD": "{^hold}",
-  "HR*EPBT": "{^ulent}",
-  "HR*ET": "{^let}",
-  "HR*EUPBS": "{^liness}",
-  "HR-L": "{^logical}",
-  "HR-LG": "{^ological}",
-  "HRA*EBG": "{^like}",
-  "HRA*EURT": "{^ularity}",
-  "HRA*PBD": "{^land}",
-  "HRA/TOER": "{^ulatory}",
-  "HRAER": "{^ulary}",
-  "HRAEUGS": "{^ulation}",
-  "HRAEUL/KWRA": "{^lalia}",
-  "HRAEUT/-D": "{^lated}",
-  "HRAEUT/-G": "{^ulating}",
-  "HRAO*EUF": "{^life}",
-  "HRAOEUT/EUBG": "{^lytic}",
-  "HRAOEUZ": "{^alize}",
-  "HRAOEUZ/-D": "{^alized}",
-  "HRAPBT": "{^lant}",
-  "HRAR": "{^ular}",
-  "HRARL": "{^ularly}",
-  "HRER": "{^ler}",
-  "HREU": "{^ly}",
-  "HREUPBG": "{^ling}",
-  "HREURBS": "{^licious}",
-  "HREUT/EUBG": "{^lytic}",
-  "HRO*": "{^ol}",
-  "HRO*PB": "{^ulon}",
-  "HROP/THEU": "{^ulopathy}",
-  "HROP/THEUS": "{^ulopathies}",
-  "HRUS": "{^ulus}",
-  "HUFRT": "{^hurst}",
-  "HUPB/HUPB": "{^00}",
-  "HUPBZ": "{^00}",
-  "K*EUPB": "{^kin}",
-  "K*L": "{^ical}",
-  "K-L": "{^cal}",
-  "K-LS": "{^icals}",
-  "KA*EUR": "{^care}",
-  "KA*EUS": "{^case}",
-  "KA/PAEUD": "{^capade}",
-  "KAE": "{^key}",
-  "KAEUGS": "{^ication}",
-  "KAEUT": "{^cate}",
-  "KAO*E/-P": "{^keep}",
-  "KAO*EP/*ER": "{^keeper}",
-  "KAO*EP/-G": "{^keeping}",
-  "KAOEUPB": "{^kine}",
-  "KHAOEZ/KWRA": "{^chezia}",
-  "KHEU/TPHEU": "{^cini}",
-  "KHRAFT": "{^clast}",
-  "KHREU": "{^ically}",
-  "KHREURBGS": "{^ically}{,}",
-  "KHROR/AOEPL/EUBG": "{^chloremic}",
-  "KO*BGS": "{^cox}",
-  "KOBG/A*L": "{^coccal}",
-  "KOBG/KWRUS": "{^coccus}",
-  "KOBG/US": "{^coccus}",
-  "KRA*T": "{^crat}",
-  "KRAES": "{^cracy}",
-  "KRAO*EUD": "{^icide}",
-  "KRAO*EUT/OES": "{^cytose}",
-  "KRAOEUT/-S": "{^cytes}",
-  "KRAOEUT/EUBG": "{^cytic}",
-  "KRAOEUT/OES": "{^cytose}",
-  "KRAOEUT/OT/EUBG": "{^cytotic}",
-  "KRAOEUT/PAOEPB/KWRA": "{^cytopenia}",
-  "KRAOEUT/PAOEPB/KWRARBGS": "{^cytopenia}",
-  "KRAT": "{^ocrat}",
-  "KWAOEUPB": "{^quine}",
-  "KWR*US": "{^eus}",
-  "KWRA": "{^a}",
-  "KWRA*": "{^ia}",
-  "KWRA*E": "{^ay}",
-  "KWRA*PB": "{^ean}",
-  "KWRA/HOL/EUBG": "{^aholic}",
-  "KWRAE": "{^ee}",
-  "KWRAER": "{^iary}",
-  "KWRAEUGS": "{^iation}",
-  "KWRAEUT": "{^iate}",
-  "KWRAFP": "{^ia}{,}",
-  "KWRAL": "{^ial}",
-  "KWRALT": "{^iality}",
-  "KWRAOE": "{^ey}",
-  "KWRAOEUL": "{^ile}",
-  "KWRAPB": "{^ian}",
-  "KWRARBG": "{^iarch}",
-  "KWRARBGS": "{^ia}{,}",
-  "KWRAS": "{^ias}",
-  "KWRAT/EUBG": "{^iatic}",
-  "KWRAT/EUF": "{^iative}",
-  "KWREBGT/AT/EUBG": "{^iectatic}",
-  "KWREBGT/EUBG": "{^iectic}",
-  "KWREFT": "{^iest}",
-  "KWREL/HRA": "{^iella}",
-  "KWRER": "{^ier}",
-  "KWREU": "{^y}",
-  "KWREU/-S": "{^ies}",
-  "KWREUD": "{^ied}",
-  "KWREUFP": "{^y}{.}",
-  "KWREUFPL": "{^y}{.}",
-  "KWREUFPLT": "{^y}",
-  "KWREUS": "{^ies}",
-  "KWRO": "{^o}",
-  "KWRO*/TKPWRAPL": "{^iogram}",
-  "KWRO*E": "{^io}",
-  "KWRO*PBD": "{^ioned}",
-  "KWRO*US": "{^eous}",
-  "KWROE/PHA": "{^ioma}",
-  "KWROEL": "{^iole}",
-  "KWROEPB/KWREU": "{^ioni}",
-  "KWROG/TPEU": "{^iography}",
-  "KWROP/THEU": "{^eopathy}",
-  "KWROPB": "{^ion}",
-  "KWROPL": "{^iom}",
-  "KWROR": "{^ior}",
-  "KWRORL": "{^iorly}",
-  "KWRORT": "{^iority}",
-  "KWROT/PHEU": "{^iotomy}",
-  "KWROUS": "{^ious}",
-  "O*EFR": "{^over}",
-  "O*EFRS": "{^overs}",
-  "O*EPB": "{^one}",
-  "O*EPL": "{^onomy}",
-  "O*EPT": "{^opathy}",
-  "O*ES": "{^ose}",
-  "O*EUL": "{^ile}",
-  "O*EUS": "{^osis}",
-  "O*L": "{^ol}",
-  "O*PB": "{^on}",
-  "O*PBS": "{^ons}",
-  "O*R": "{^or}",
-  "O*RS": "{^ors}",
-  "O*T": "{^otte}",
-  "O*ULS": "{^ously}",
-  "OBG/A*L": "{^occal}",
-  "OBGS/*EUPB": "{^oxin}",
-  "OBGS/KWRA": "{^oxia}",
-  "OE/PHA": "{^oma}",
-  "OE/RA/PHA": "{^orama}",
-  "OE/SA": "{^osa}",
-  "OEL/SA*EUGS": "{^olization}",
-  "OEP/EUBG": "{^opic}",
-  "OEP/KWRA": "{^opia}",
-  "OEPB/KWRUPL": "{^onium}",
-  "OEPL/KWRA": "{^oma}",
-  "OES": "{^ose}",
-  "OES/EUS": "{^osis}",
-  "OES/KWRUPL": "{^osum}",
-  "OES/SKWRUS": "{^osus}",
-  "OES/US": "{^osus}",
-  "OEUBG": "{^oic}",
-  "OEUD": "{^oid}",
-  "OFBG/PEU": "{^oscopy}",
-  "OFT/PHEU": "{^ostomy}",
-  "OG/TPEU": "{^ography}",
-  "OL/SEUS": "{^olysis}",
-  "OLG": "{^ology}",
-  "OLG/EUFT": "{^ologist}",
-  "OP/HREUS": "{^opolis}",
-  "OP/THEU": "{^opathy}",
-  "OPB/EUBG": "{^onic}",
-  "OPBLG/K-L": "{^ogical}",
-  "OPL/EUBG": "{^omic}",
-  "OPL/TREU": "{^ometry}",
-  "OPLT/*ER": "{^ometer}",
-  "OPS/EUS": "{^opsis}",
-  "OR/KWROUS": "{^orious}",
-  "ORPL": "{^orium}",
-  "OS": "{^os}",
-  "OS/KO/PEU": "{^oscopy}",
-  "OS/SKWRUS": "{^osus}",
-  "OT/EUBG": "{^otic}",
-  "OT/PHEU": "{^otomy}",
-  "OUS": "{^ous}",
-  "P*EU/TPH*EU": "{^pini}",
-  "P-D": "{^ped}",
-  "PA*EBGT": "{^pathic}",
-  "PA*ET": "{^pathy}",
-  "PA*ET/-S": "{^pathies}",
-  "PA*S": "{^pass}",
-  "PAOED/EUBG": "{^pedic}",
-  "PAOED/KWRA": "{^pedia}",
-  "PAOEPB/EUBG": "{^penic}",
-  "PAOEPB/KWRA": "{^penia}",
-  "PAOEPB/KWRAFPLT": "{^penia}{.}",
-  "PAOEPB/KWRARBGS": "{^penia}{,}",
-  "PEPB/EPL": "{^penem}",
-  "PEPBT/TPHOEUBG": "{^pentaenoic}",
-  "PH*/*EU/TK*E": "{^mide}",
-  "PH*EPB": "{^men}",
-  "PHA*ED": "{^made}",
-  "PHA*EUBG": "{^make}",
-  "PHA*EUG": "{^making}",
-  "PHA*EURBG": "{^maker}",
-  "PHA*PB": "{^man}",
-  "PHAB": "{^mab}",
-  "PHAEUGS": "{^mation}",
-  "PHAO*EPB": "{^amine}",
-  "PHAO*ERT": "{^meter}",
-  "PHAO*EUD": "{^mide}",
-  "PHAO*EUPB": "{^amine}",
-  "PHAOES/KWRUPL": "{^mysium}",
-  "PHAOEUFG": "{^imizing}",
-  "PHAOEUFPB": "{^mycin}",
-  "PHAOEUS/*EUPB": "{^mycin}",
-  "PHAPB/SH-P": "{^manship}",
-  "PHAT/EUBG": "{^matic}",
-  "PHEG/HREU": "{^megaly}",
-  "PHER/-S": "{^mers}",
-  "PHET/REU": "{^metry}",
-  "PHO*EFT": "{^most}",
-  "PHO*PBD": "{^mond}",
-  "PHOEPBL": "{^monal}",
-  "PHOPBG/*ER": "{^monger}",
-  "PHOPBT": "{^mont}",
-  "PHRAEFT": "{^plasty}",
-  "PHRAEU/PHRA*EU/SHA": "{^plasia}",
-  "PHRAEU/SHA": "{^plasia}",
-  "PHRAEURB/KWRA": "{^plasia}",
-  "PHRAEUS/EUBG": "{^plasic}",
-  "PHRAEUS/KWRA": "{^plasia}",
-  "PHRAFT": "{^plast}",
-  "PHRAFT/KWREU": "{^plasty}",
-  "PHRAOEPBLG/EUBG": "{^plegic}",
-  "PHRAOEPBLG/KWRA": "{^plegia}",
-  "PHROEUFT": "{^oplasty}",
-  "PO*RT": "{^port}",
-  "POEUS/EUS": "{^poiesis}",
-  "PRABGS/KWRA": "{^praxia}",
-  "PRO*EUPB": "{^protein}",
-  "PROES/EUS": "{^porosis}",
-  "PW*UR/KWREU": "{^bury}",
-  "PW*URG": "{^burg}",
-  "PWA*BG": "{^back}",
-  "PWA*PBD": "{^band}",
-  "PWA*RBGS": "{^back}{,}",
-  "PWAO*BG": "{^book}",
-  "PWAO*RD": "{^board}",
-  "PWAOEUT/EUBG": "{^biotic}",
-  "PWERG": "{^berg}",
-  "PWEUGS": "{^ibition}",
-  "PWEUL/*EUPB": "{^bilin}",
-  "PWHRAOUPL": "{^blume}",
-  "PWHRER": "{^bler}",
-  "PWHREU": "{^ably}",
-  "PWO*LG": "{^bolic}",
-  "PWO*UPBD": "{^bound}",
-  "PWO/RO": "{^boro}",
-  "PWRO*": "{^boro}",
-  "PWURG": "{^burg}",
-  "R*D": "{^rd}",
-  "R*EUBG": "{^ric}",
-  "R-D": "{^red}",
-  "RA*EUGS": "{^ration}",
-  "RAEBL": "{^rably}",
-  "RAL": "{^ural}",
-  "RAO*EUFD": "{^arized}",
-  "RAO*PL": "{^room}",
-  "RAOUB/SEUPB": "{^rubicin}",
-  "RAPBS": "{^erance}",
-  "RAT/EUF": "{^arative}",
-  "RES": "{^ress}",
-  "REU": "{^ry}",
-  "REUS": "{^aries}",
-  "RO/PHAOEUD": "{^romide}",
-  "ROES/ES": "{^roses}",
-  "ROUS": "{^rous}",
-  "S*E": "{^es}",
-  "S*EPT": "{^sept}",
-  "S*ER": "{^zer}",
-  "S*EU/TPH*EU": "{^zini}",
-  "S*EUPBL": "{^icillin}",
-  "SA*EFT": "{^icist}",
-  "SA*EUGS": "{^ization}",
-  "SABG/RAOEUD": "{^saccharide}",
-  "SAEUGS": "{^sation}",
-  "SAO*EUD": "{^icide}",
-  "SAOEPL/KWRA": "{^cemia}",
-  "SAOEUBG/HRAO*EUPB": "{^cyline}",
-  "SAOEUD/A*L": "{^cidal}",
-  "SEFL": "{^self}",
-  "SEFLS": "{^selves}",
-  "SEPT/O*R": "{^ceptor}",
-  "SEU": "{^cy}",
-  "SEU/TO*": "{^cito}",
-  "SEUL/*EUPB": "{^cillin}",
-  "SEUT/EUBG": "{^cytic}",
-  "SH*UPB": "{^tion}",
-  "SH-P": "{^ship}",
-  "SH-PS": "{^ships}",
-  "SHA*U": "{^shaw}",
-  "SHAO*EUPB": "{^shine}",
-  "SHEUR": "{^shire}",
-  "SHUPL": "{^tium}",
-  "SK*EU": "{^ski}",
-  "SKAEUP": "{^scape}",
-  "SKO*EP": "{^oscopy}",
-  "SKO*EUP": "{^oscope}",
-  "SKOP/EUBG": "{^scopic}",
-  "SKOP/KHREU": "{^scopically}",
-  "SKWR*E": "{^e}",
-  "SKWRUS": "{^us}",
-  "SKWRA": "{^a}",
-  "SKWRAO": "{^oo}",
-  "SKWRAOEU": "{^ae}",
-  "SKWRAOEUL": "{^ile}",
-  "SKWRE": "{^e}",
-  "SKWREFR": "{^ever}",
-  "SKWREPB/*EUFT": "{^genicity}",
-  "SKWREPB/EUBG": "{^genic}",
-  "SKWREPB/SEUS/TEU": "{^genicity}",
-  "SKWREU": "{^i}",
-  "SKWREUS": "{^ies}",
-  "SKWRO": "{^o}",
-  "SKWRO*/KRAOEUT": "{^ocyte}",
-  "SKWRO/KRAOEUT": "{^ocyte}",
-  "SKWRO/RA/PHA": "{^orama}",
-  "SKWRO/SA": "{^osa}",
-  "SKWRO/TKPWOG": "{^ogogue}",
-  "SKWROF": "{^off}",
-  "SKWROUPB": "{^down}",
-  "SKWROUT": "{^out}",
-  "SKWRU": "{^u}",
-  "SKWRUP": "{^up}",
-  "SKWRUPBD": "{^under}",
-  "SKWRUPL": "{^um}",
-  "SO*EPL": "{^some}",
-  "SO*EUD": "{^azoid}",
-  "SO*EUT": "{^zoite}",
-  "SO*PB": "{^son}",
-  "SO*PL": "{^some}",
-  "SPOR/*EUPB": "{^sporin}",
-  "SREUFP": "{^ovich}",
-  "SREUL": "{^ville}",
-  "SROR": "{^vore}",
-  "ST-BG": "{^istic}",
-  "STA*PBD": "{^stand}",
-  "STAO*EPB": "{^stein}",
-  "STAO*EUPB": "{^stein}",
-  "STER": "{^ster}",
-  "STH-S": "{^s}{?}",
-  "STO*EPB": "{^stone}",
-  "STO*PB": "{^ston}",
-  "STOPB": "{^ston}",
-  "STP-S": "{^s}{?}",
-  "STPA": "{^a}",
-  "STPE": "{^e}",
-  "STPEU": "{^i}",
-  "STPH-D": "{^ed}{?}",
-  "STPH-S": "{^s}{?}",
-  "STPO": "{^o}",
-  "STPU": "{^u}",
-  "STR/EUBG": "{^centric}",
-  "T*EUF": "{^tive}",
-  "T*EUFT": "{^tivity}",
-  "TA*BGT": "{^tatic}",
-  "TA*EURBG": "{^taker}",
-  "TAEUGS": "{^tation}",
-  "TAEURPB": "{^itarian}",
-  "TAFT/EUBG": "{^tastic}",
-  "TAO*EUPL": "{^time}",
-  "TAO*EUT": "{^tight}",
-  "TAO*EUZ": "{^atize}",
-  "TAOEUT/EUS": "{^titis}",
-  "TAOEUZ/AEUGS": "{^tization}",
-  "TAOUD": "{^itude}",
-  "TAOUR": "{^ture}",
-  "TAOUZ/PHAB": "{^tuzumab}",
-  "TAPBLG": "{^itage}",
-  "TAS/TEUBG": "{^tastic}",
-  "TEU": "{^ity}",
-  "TEU/TPHEU": "{^tini}",
-  "TEUF": "{^tive}",
-  "TEUFL": "{^itively}",
-  "TEUGS": "{^etition}",
-  "TEURBS": "{^ticious}",
-  "TEUS/TEU": "{^ticity}",
-  "THAL/*ER": "{^thaler}",
-  "THAL/PHEU/KWRA": "{^ophthalmia}",
-  "THO*EUPB": "{^000}",
-  "THR*U": "{^through}",
-  "TK*EPB": "{^den}",
-  "TK-D": "{^ded}",
-  "TK-RBS": "{^s} --", // FIXME: might not match affixList regex
-  "TKA*EL": "{^dale}",
-  "TKAEUGS": "{^dation}",
-  "TKEU/TOE": "{^dito}",
-  "TKEU/TPHEU": "{^dini}",
-  "TKEUGS": "{^dition}",
-  "TKPWAFPL": "{^gasm}",
-  "TKPWHRO*EPB": "{^globin}",
-  "TKPWRAF/KWREU": "{^graphy}",
-  "TKPWRAO*UP": "{^group}",
-  "TKPWRO*UPBD": "{^ground}",
-  "TKRA*UPB": "{^drawn}",
-  "TKROEPL": "{^drome}",
-  "TO*P": "{^top}",
-  "TO*PB": "{^ton}",
-  "TO*UPB": "{^town}",
-  "TOEP": "{^tope}",
-  "TOER": "{^tory}",
-  "TOES/EUS": "{^tosis}",
-  "TOG/TPEU": "{^tography}",
-  "TOLG": "{^otology}",
-  "TOS/KO/PEU": "{^toscopy}",
-  "TOT/PHEU": "{^totomy}",
-  "TOUS": "{^tous}",
-  "TP*EUL/EUBG": "{^philic}",
-  "TP*EULG": "{^philic}",
-  "TP*EUPB": "{^fin}",
-  "TP-PLD": "{^ed}{.}",
-  "TP-PLS": "{^s}{.}",
-  "TPAEUBGS": "{^ification}",
-  "TPAEUPBLG/KWRA": "{^phagia}",
-  "TPAO*D": "{^food}",
-  "TPAO*ELD": "{^field}",
-  "TPAO*EUL": "{^phile}",
-  "TPAO*EUR": "{^ifier}",
-  "TPAOEL/KWRA": "{^philia}",
-  "TPAOEU": "{^ify}",
-  "TPAOEU/-D": "{^ified}",
-  "TPAOEU/-S": "{^ifies}",
-  "TPAOEUD": "{^ified}",
-  "TPAOEUDZ": "{^ifieds}",
-  "TPAOEUG": "{^ifying}",
-  "TPEUBG/A*EUGS": "{^fication}",
-  "TPEUL/EUBG": "{^philic}",
-  "TPEUL/EUS/TEU": "{^philicity}",
-  "TPEUL/KWRA": "{^philia}",
-  "TPEUL/KWRAS": "{^philias}",
-  "TPEULS/TEU": "{^philicity}",
-  "TPHA*EUT": "{^inate}",
-  "TPHAEL": "{^inally}",
-  "TPHAEUGS": "{^ination}",
-  "TPHAEUT/O*R": "{^inator}",
-  "TPHAL": "{^inal}",
-  "TPHAO*E": "{^ni}",
-  "TPHAO*ES": "{^nese}",
-  "TPHAO*EUPB": "{^inine}",
-  "TPHAOEUZ": "{^nize}",
-  "TPHAOEUZ/-G": "{^nizing}",
-  "TPHER": "{^ener}",
-  "TPHERPBLG/EUBG": "{^nergic}",
-  "TPHEUFT": "{^ist}",
-  "TPHEUPBG": "{^ening}",
-  "TPHO*US": "{^inous}",
-  "TPHOFT/PHEU": "{^anostomy}",
-  "TPHOS/KO/PEU": "{^noscopy}",
-  "TPHOUS": "{^inous}",
-  "TPHR*EBGS": "{^flection}",
-  "TPHRABGS/EUS": "{^phylaxis}",
-  "TPHREU": "{^fully}",
-  "TPHREUFT": "{^philicity}",
-  "TPHROBGS/SEUPB": "{^floxacin}",
-  "TPHROPBGS/SEUPB": "{^floxacin}",
-  "TPO*ELD": "{^fold}",
-  "TPO*RPL": "{^form}",
-  "TPO*UPBD": "{^found}",
-  "TPOEB/EUS/TEU": "{^phobicity}",
-  "TPOR/AOES/EUS/TKOS": "{^pheresis}",
-  "TPOR/EUBG": "{^phoric}",
-  "TPR-Z": "{^s} from", // FIXME: might not match affixList regex
-  "TPRAOES/EUS": "{^pheresis}",
-  "TPRO*PBT": "{^front}",
-  "TR*EU/TPHOEUBG": "{^trienoic}",
-  "TRAOUS/EUF": "{^trusive}",
-  "TRO*EF": "{^trophy}",
-  "TRO*EP": "{^tropy}",
-  "TROEF/EUBG": "{^trophic}",
-  "TROEP/*EUBG": "{^tropic}",
-  "TROEP/*EUPB": "{^tropin}",
-  "TROEP/EUBG": "{^tropic}",
-  "TROEP/KWREU": "{^tropy}",
-  "TROEP/KWRUPL": "{^tropium}",
-  "TROEPG": "{^otropic}",
-  "TROEUPG": "{^otropic}",
-  "TROF/EUBG": "{^trophic}",
-  "TRUPL": "{^trum}",
-  "WA*E": "{^way}",
-  "WA*ER": "{^wear}",
-  "WA*ES": "{^ways}",
-  "WA*EUR": "{^ware}",
-  "WA*EUT": "{^uate}",
-  "WA*RD": "{^ward}",
-  "WAEL": "{^ually}",
-  "WAEUGS": "{^uation}",
-  "WAL": "{^ual}",
-  "WAPBS": "{^uance}",
-  "WO*PL": "{^woman}",
-  "WO*RBG": "{^work}",
-  "WO*RD": "{^word}",
-  "WO*RLD": "{^world}",
-  "WO*RPL": "{^worm}",
-  "WOS/TEU": "{^uosity}",
-  "WOUS": "{^uous}",
-  "WUPL": "{^uum}",
-  "*EPBLG": "{^-edge}",
-  "A*EURBG": "{^-acre}",
-  "A*UR/KWREPBT/-D": "{^-oriented}",
-  "A/KWAO*EURD": "{^-acquired}",
-  "A/PRO*FD": "{^-approved}",
-  "ARBGZ": "{^-A}",
-  "ERBGZ": "{^-E}",
-  "EUPB/TKAO*UFD": "{^-induced}",
-  "EURBGZ": "{^-I}",
-  "H*ED": "{^-head}",
-  "H-PB/KA/TPHAEUD/KWRAPB": "{^-Canadian}",
-  "H-PB/KAPB/TKA": "{^-Canada}",
-  "H-PB/PHERPB": "{^-American}",
-  "H-RBGZ": "{^-H}",
-  "HO*UR": "{^-hour}",
-  "HR-RBGZ": "{^-L}",
-  "HRA*EPB": "{^-lane}",
-  "HRAO*BG": "{^-look}",
-  "HRO*PBG": "{^-long}",
-  "K*UP": "{^-cup}",
-  "K-RBGZ": "{^-K}",
-  "KA*EURT": "{^-carat}",
-  "KAO*BGD": "{^-cooked}",
-  "KHRO*PLT": "{^-kilometer}",
-  "KHRO*R": "{^-color}",
-  "KP-RBGZ": "{^-X}",
-  "KPHRAO*ET": "{^-complete}",
-  "KR-RBGZ": "{^-C}",
-  "KRO*ELD": "{^-controlled}",
-  "KW-RBGZ": "{^-Q}",
-  "KWA*ERT": "{^-quart}",
-  "KWA*LT": "{^-quality}",
-  "KWR-RBGZ": "{^-Y}",
-  "KWRA*RD": "{^-yard}",
-  "KWRAO*ER": "{^-year}",
-  "KWRAUR/KWREPBT": "{^-orient}",
-  "O*D": "{^-odd}",
-  "O*ELD": "{^-old}",
-  "O*EPBD": "{^-owned}",
-  "O*F": "{^-off}",
-  "O*UPBS": "{^-ounce}",
-  "O*UT": "{^-out}",
-  "ORBGZ": "{^-O}",
-  "P*ERPB": "{^-person}",
-  "P*URP": "{^-purpose}",
-  "P-RBGZ": "{^-P}",
-  "PA*EUPBLG": "{^-page}",
-  "PA*EURBT": "{^-patient}",
-  "PA*RT": "{^-part}",
-  "PAO*ES": "{^-piece}",
-  "PH*EUPBT": "{^-minute}",
-  "PH-RBGZ": "{^-M}",
-  "PHAO*ED/KWRAEUT/-D": "{^-mediated}",
-  "PHAO*ED/KWRAEUTD": "{^-mediated}",
-  "PHAO*EUL": "{^-mile}",
-  "PHO*EPBT": "{^-month}",
-  "PHO*UPBT/-D": "{^-mounted}",
-  "PHRAO*EU": "{^-ply}",
-  "PO*UPBD": "{^-pound}",
-  "PRAO*EUS": "{^-price}",
-  "PRAO*F": "{^-proof}",
-  "PRO*EPT": "{^-appropriate}",
-  "PRO*ET": "{^-appropriate}",
-  "PW*ED": "{^-bed}",
-  "PW*EULT": "{^-built}",
-  "PW*EUT": "{^-bit}",
-  "PW*PL": "{^-bedroom}",
-  "PW*URPBG": "{^-burning}",
-  "PW-RBGZ": "{^-B}",
-  "PWA*ERPBG": "{^-bearing}",
-  "PWA*EUFD": "{^-based}",
-  "PWA*EUS": "{^-base}",
-  "PWR*PL": "{^-bathroom}",
-  "R*EPLTD": "{^-related}",
-  "R*EUFPBT": "{^-resistant}",
-  "R*EUPBT": "{^-resistant}",
-  "R-RBGZ": "{^-R}",
-  "RA*ELT/-D": "{^-related}",
-  "RA*ELTD": "{^-related}",
-  "RO*EFT/-D": "{^-roasted}",
-  "S*EBGD": "{^-second}",
-  "S*ERD": "{^-certified}",
-  "S*G": "{^-something}",
-  "S-RBGZ": "{^-S}",
-  "SA*EUF": "{^-safe}",
-  "SA*F": "{^-sav}",
-  "SA*FG": "{^-saving}",
-  "SAO*EUD/-D": "{^-sided}",
-  "SAO*EUFD": "{^-sized}",
-  "SAO*EUZ": "{^-size}",
-  "SHA*EUP": "{^-shape}",
-  "SHA*EUPD": "{^-shaped}",
-  "SK*EBL": "{^-susceptible}",
-  "SKA*PB": "{^-scan}",
-  "SKWR*EPBD": "{^-end}",
-  "SKWR*EUPB": "{^-in}",
-  "SKWR*EUPB/SPAOEUR/-D": "{^-inspired}",
-  "SKWR*EUPB/SPAOEUR/-G": "{^-inspiring}",
-  "SKWR*UP": "{^-up}",
-  "SKWR*UP/*ER": "{^-upper}",
-  "SKWR*UPS": "{^-ups}",
-  "SKWR-RBGZ": "{^-J}",
-  "SKWRA*S": "{^-ass}",
-  "SKWRAO*BG": "{^-look}",
-  "SKWRAO*EUBG": "{^-like}",
-  "SKWRAOBG": "{^-look}",
-  "SKWRAOEUBG": "{^-like}",
-  "SKWRAOG": "{^-looking}",
-  "SKWRAUL": "{^-all}",
-  "SKWREUPB/SPAOEUR/-D": "{^-inspired}",
-  "SKWRO*F": "{^-off}",
-  "SKWRO*FS": "{^-offs}",
-  "SKWRO*PB": "{^-on}",
-  "SKWRO*UPB": "{^-down}",
-  "SKWRO*UPBS": "{^-downs}",
-  "SKWRO*UT": "{^-out}",
-  "SO*ERBT/-D": "{^-associated}",
-  "SO*ERBTD": "{^-associated}",
-  "SP*EFBG": "{^-specific}",
-  "SP*EUFBG": "{^-specific}",
-  "SPAO*ED": "{^-speed}",
-  "SPAO*EG": "{^-speaking}",
-  "SPAO*ERPBS/-D": "{^-experienced}",
-  "SR-RBGZ": "{^-V}",
-  "SRO*ELT": "{^-volt}",
-  "ST*EP": "{^-step}",
-  "STA*EUPBLG": "{^-stage}",
-  "STAO*EUL": "{^-style}",
-  "STK-RBGZ": "{^-Z}",
-  "STO*ER": "{^-story}",
-  "STR*D": "{^-centered}",
-  "STR*EPBG": "{^-strength}",
-  "T*ERPL": "{^-term}",
-  "T-RBGZ": "{^-T}",
-  "TA*EPL": "{^-time}",
-  "TA*EUFT": "{^-taste}",
-  "TA*UL": "{^-tall}",
-  "TAO*ER": "{^-tier}",
-  "TAO*EUP": "{^-type}",
-  "TAO*EUPL/*ER": "{^-timer}",
-  "TH*EUBG": "{^-thick}",
-  "THR*EFPBG": "{^-threatening}",
-  "TK*EG": "{^-degree}",
-  "TK-RBGZ": "{^-D}",
-  "TKA*EU": "{^-day}",
-  "TKAO*EPBT": "{^-dependent}",
-  "TKAO*R": "{^-door}",
-  "TKHRA*R": "{^-dollar}",
-  "TKO*UPB": "{^-down}",
-  "TKPHEPBL": "{^-dimensional}",
-  "TKPW-RBGZ": "{^-G}",
-  "TKPWA*EUPBLG": "{^-gauge}",
-  "TKPWHRO*PB": "{^-gallon}",
-  "TKPWR*EUT": "{^-grit}",
-  "TP*EURBT": "{^-efficient}",
-  "TP-RBGZ": "{^-F}",
-  "TPA*EUPL": "{^-fame}",
-  "TPA*EUPL/OUS": "{^-famous}",
-  "TPA*EUS/-G": "{^-facing}",
-  "TPAO*T": "{^-foot}",
-  "TPH*EBGT/-D": "{^-infected}",
-  "TPH*EBGTD": "{^-infected}",
-  "TPH*FP": "{^-inch}",
-  "TPH-RBGZ": "{^-N}",
-  "TPHAO*EF": "{^-naive}",
-  "TPHO*EPB": "{^-known}",
-  "TPHRA*EUFR": "{^-flavor}",
-  "TPO*EUBGS/-D": "{^-focused}",
-  "TPR*EPBD/HREU": "{^-friendly}",
-  "TPRAO*E": "{^-free}",
-  "TPRAO*EUD": "{^-fried}",
-  "TRAO*ETD": "{^-treated}",
-  "URBGZ": "{^-U}",
-  "W*EPBS": "{^-influence}",
-  "W-RBGZ": "{^-W}",
-  "WAO*EBG": "{^-week}",
-  "WAO*EUD": "{^-wide}",
-  "WAO*EUS": "{^-wise}",
-  "ES": "{^es}",
-  // ["/*PBS", "iness"],
-  // ["/*S", "s"],
-  // ["/-SZ", "s"],
-  // ["/-Z", "s"],
-  // ["/AEUR/TEU", "arity"],
-  // ["/AL/TEU", "ality"],
-  // ["/EBG/HREU", "ically"],
-  // ["/EUL/TEU", "ility"],
-  // ["/HREUPBS", "liness"],
-  // ["/KAO*ERP", "keeper"],
-  // ["/KWRA*U", "ier"],
-  // ["/KWRUS", "us"],
-  // ["/O*US", "us"],
-  // ["/OPL/TER", "ometer"],
-  // ["/PHA*EUBG/-G", "making"],
-  // ["/PHAO*EUFG", "imizing"],
-  // ["/PHO*PBG/*ER", "monger"],
-  // ["/SAO*EUD/A*L", "cidal"],
-  // ["/SKWR*US", "us"],
-  // ["/SKWRAOED", "ed"],
-  // ["/TEUFT", "ticity"],
-  // ["/TPHAER", "ary"],
-  // ["/TPHALT", "ality"],
-};
-
-const personalDictionaries: PersonalDictionaryNameAndContents[] = [
-  // [ "test-typey-type.json", testTypeyTypeDict],
-  [ "test-emoji.json", testEmojiDict],
-  [ "test-ruby.json", testRubyDict],
-  [ "test-react.json", testReactDict],
-  // [ "test-plover.json", testPloverDict],
-  [ "test-prefixes.json", testPrefixesDict],
-  [ "test-suffixes.json", testSuffixesDict],
-  [ "test-aussie.json", testAussieDict],
-];
+import {
+  testTypeyTypeDict,
+  testPloverDict,
+  personalDictionaries,
+} from "./createStrokeHintForPhrase.fixtures";
 
 let sortedAndCombinedLookupDictionary = createAGlobalLookupDictionary(personalDictionaries, testTypeyTypeDict, testPloverDict);
